#1044b: split pkg/daemon/daemon.go (2,851 → 346 LOC)
daemon.go held 23 functions across 2,851 LOC. Two functions dominated:
Run (~942 LOC, the main daemon event loop) and applyConfigLocked
(~870 LOC, the post-commit reconciliation pipeline).

Extracted into focused subfiles, including each function's preceding
doc comment so go doc still associates them correctly:
- daemon_apply.go (985 LOC, 6 fns): bootstrapFromFile, applyConfig,
  commitAndApply, syncAndApply, commitConfirmedAndApply, applyConfigLocked
- daemon_run.go (1,153 LOC, 4 fns): collectAppliedTunnels, Run,
  enableForwarding, inferIPv6StaticNextHopInterfaces
- daemon_dhcp.go (163 LOC, 5 fns): startDHCPClients,
  dhcpLeaseChangeRequiresRecompile, resolveJunosIfName,
  resolveConfigSubnetLinuxName, stripCIDR
- daemon_ra.go (110 LOC, 2 fns): buildRAConfigs, cloneRAInterfaceConfig
- daemon_health.go (73 LOC, 5 fns): recordCompileFailure,
  recordCompileSuccess, CompileHealthSnapshot,
  shouldScheduleStandbyNeighborRefresh, scheduleStandbyNeighborRefresh

After split:
- daemon.go: 2,851 → 346 LOC (88% reduction; kept: types, New)
- 5 new files

Splitting Run or applyConfigLocked themselves requires per-stage
design and is deferred to follow-up PRs.

`goimports -w` to prune per-file imports. `go build ./...` clean.
`go test ./pkg/daemon/...` pass.
