+++ pkg/daemon/daemon.go
@@ -2,36 +2,24 @@
 package daemon
 
 import (
-	"bytes"
 	"context"
-	"encoding/binary"
 	"fmt"
 	"log/slog"
 	"net"
 	"os"
-	"os/exec"
-	"os/signal"
-	"path/filepath"
-	"slices"
-	"sort"
-	"strconv"
 	"strings"
 	"sync"
 	"sync/atomic"
-	"syscall"
 	"time"
 
 	"github.com/vishvananda/netlink"
 	"golang.org/x/sync/semaphore"
 
-	"github.com/psaab/xpf/pkg/api"
-	"github.com/psaab/xpf/pkg/cli"
 	"github.com/psaab/xpf/pkg/cluster"
 	"github.com/psaab/xpf/pkg/config"
 	"github.com/psaab/xpf/pkg/configstore"
 	"github.com/psaab/xpf/pkg/conntrack"
 	"github.com/psaab/xpf/pkg/dataplane"
-	dpuserspace "github.com/psaab/xpf/pkg/dataplane/userspace"
 	"github.com/psaab/xpf/pkg/dhcp"
 	"github.com/psaab/xpf/pkg/dhcprelay"
 	"github.com/psaab/xpf/pkg/dhcpserver"
@@ -39,7 +27,6 @@ import (
 	"github.com/psaab/xpf/pkg/feeds"
 	"github.com/psaab/xpf/pkg/flowexport"
 	"github.com/psaab/xpf/pkg/frr"
-	"github.com/psaab/xpf/pkg/fwdstatus"
 	"github.com/psaab/xpf/pkg/grpcapi"
 	"github.com/psaab/xpf/pkg/ipsec"
 	"github.com/psaab/xpf/pkg/lldp"
@@ -303,85 +290,7 @@ type CompileHealth struct {
 	LastErrorUnixSec int64
 }
 
-// recordCompileFailure tracks a dataplane compile failure and emits an
-// escalating log (#758). The first failure remains a single WARN;
-// every Nth repeat re-emits at ERROR level so an operator tailing the
-// journal sees the degraded state without needing to know the original
-// failure text. A success clears the counter via recordCompileSuccess.
-func (d *Daemon) recordCompileFailure(err error) {
-	d.compileHealthMu.Lock()
-	d.compileFailureCount++
-	d.compileLastError = err.Error()
-	d.compileLastErrorUnixSec = time.Now().Unix()
-	count := d.compileFailureCount
-	everOk := d.compileEverSucceeded
-	d.compileHealthMu.Unlock()
-
-	// First WARN fires on every failure — matches pre-#758 behaviour.
-	// Escalate to ERROR on the 5th consecutive failure with no prior
-	// success, and again every 10 failures thereafter, so a persistent
-	// degraded state stays visible in the log without flooding.
-	slog.Warn("failed to compile dataplane", "err", err, "attempt", count, "ever_ok", everOk)
-	if !everOk && (count == 5 || (count > 5 && count%10 == 0)) {
-		slog.Error("dataplane compile has failed repeatedly; forwarding path is degraded",
-			"attempt", count, "err", err)
-	}
-}
-
-// recordCompileSuccess clears compile failure state. The failure count
-// is intentionally preserved as a monotonic "have we ever hit this"
-// counter (exported via CompileHealthSnapshot), but the "ever ok" flag
-// flips true so /health goes back to healthy.
-func (d *Daemon) recordCompileSuccess() {
-	d.compileHealthMu.Lock()
-	d.compileEverSucceeded = true
-	d.compileLastError = ""
-	d.compileHealthMu.Unlock()
-}
-
-// CompileHealthSnapshot returns the current compile health for /health
-// and operator-facing RPCs. Safe to call concurrently.
-func (d *Daemon) CompileHealthSnapshot() CompileHealth {
-	d.compileHealthMu.Lock()
-	defer d.compileHealthMu.Unlock()
-	return CompileHealth{
-		EverSucceeded:    d.compileEverSucceeded,
-		FailureCount:     d.compileFailureCount,
-		LastError:        d.compileLastError,
-		LastErrorUnixSec: d.compileLastErrorUnixSec,
-	}
-}
-
 const standbyNeighborRefreshMinInterval = time.Second
-
-func (d *Daemon) shouldScheduleStandbyNeighborRefresh(now time.Time) bool {
-	elapsed := now.Sub(d.startTime).Nanoseconds() + 1
-	if elapsed < 1 {
-		elapsed = 1
-	}
-	last := d.lastStandbyNeighborRefresh.Load()
-	if last != 0 && elapsed >= last && elapsed-last < int64(standbyNeighborRefreshMinInterval) {
-		return false
-	}
-	return d.lastStandbyNeighborRefresh.CompareAndSwap(last, elapsed)
-}
-
-func (d *Daemon) scheduleStandbyNeighborRefresh() {
-	if d.cluster == nil || d.dp == nil {
-		return
-	}
-	cfg := d.store.ActiveConfig()
-	if cfg == nil {
-		return
-	}
-	if !d.shouldScheduleStandbyNeighborRefresh(time.Now()) {
-		return
-	}
-	go func(cfg *config.Config) {
-		d.resolveNeighborsInner(cfg, false)
-		d.maintainClusterNeighborReadiness()
-	}(cfg)
-}
 
 // New creates a new Daemon.
 func New(opts Options) *Daemon {
@@ -420,2432 +329,4 @@ func New(opts Options) *Daemon {
 		userspaceDemotionPrepUntil: make(map[int]time.Time),
 		applySem:                   semaphore.NewWeighted(1),
 	}
-}
-
-func collectAppliedTunnels(cfg *config.Config) []*config.TunnelConfig {
-	if cfg == nil {
-		return nil
-	}
-	anchorOnly := cfg.System.DataplaneType == dataplane.TypeUserspace
-	var tunnels []*config.TunnelConfig
-	for _, ifc := range cfg.Interfaces.Interfaces {
-		if ifc == nil {
-			continue
-		}
-		if ifc.Tunnel != nil && ifc.Tunnel.Source != "" {
-			tc := *ifc.Tunnel
-			tc.AnchorOnly = anchorOnly
-			tunnels = append(tunnels, &tc)
-		}
-		for _, unit := range ifc.Units {
-			if unit == nil || unit.Tunnel == nil {
-				continue
-			}
-			tc := *unit.Tunnel
-			tc.AnchorOnly = anchorOnly
-			tunnels = append(tunnels, &tc)
-		}
-	}
-	return tunnels
-}
-
-// Run starts the daemon and blocks until shutdown.
-func (d *Daemon) Run(ctx context.Context) error {
-	d.daemonCtx = ctx
-
-	// Wrap the default slog handler to support system syslog forwarding.
-	// Syslog clients are added later when config is applied.
-	d.slogHandler = logging.NewSyslogSlogHandler(slog.Default().Handler())
-	slog.SetDefault(slog.New(d.slogHandler))
-
-	slog.Info("starting xpf daemon",
-		"config", d.opts.ConfigFile,
-		"pid", os.Getpid())
-
-	// Load persisted configuration from DB, falling back to text config file
-	if err := d.store.Load(); err != nil {
-		slog.Warn("failed to load config from db", "err", err)
-	}
-
-	// If DB had no active config, bootstrap from the text config file
-	if d.store.ActiveConfig() == nil {
-		if err := d.bootstrapFromFile(); err != nil {
-			slog.Warn("failed to bootstrap config from file", "err", err)
-		}
-	} else {
-		slog.Info("configuration loaded from db")
-	}
-
-	// Enumerate PCI NICs and assign vSRX-style names (fxp0, em0, ge-X-0-Y)
-	// before any manager creation or BPF load.
-	if !d.opts.NoDataplane {
-		clusterMode := false
-		nodeID := 0
-		userspaceWorkers := 0
-		// D3 (#797): default enabled. Operators opt out via
-		// `set system dataplane rss-indirection disable`.
-		rssEnabled := true
-		var rssAllowed []string
-		// #801 Phase-B Step-0 tunables: host-scope governor + netdev
-		// budget; per-iface mlx5 coalescence. Host-scope knobs are
-		// GATED by `claim-host-tunables true` (B1). Per-iface knobs
-		// (rx-usecs/tx-usecs) follow the D3 allowlist and are applied
-		// whenever coalescence is configured.
-		var (
-			governor          string
-			netdevBudget      int
-			coalesceEnable    bool
-			coalesceRX        int
-			coalesceTX        int
-			userspaceDP       bool
-			coalesceExplicit  bool
-			claimHostTunables bool
-		)
-		if cfg := d.store.ActiveConfig(); cfg != nil {
-			if cfg.Chassis.Cluster != nil {
-				clusterMode = true
-				nodeID = cfg.Chassis.Cluster.NodeID
-			}
-			// D3 (#785): pass userspace-dp worker count so linksetup can
-			// reshape mlx5 RSS indirection before any AF_XDP bind. Zero
-			// when userspace dataplane is not in use — applyRSSIndirection
-			// treats that as a no-op.
-			if cfg.System.DataplaneType == "userspace" && cfg.System.UserspaceDataplane != nil {
-				userspaceDP = true
-				userspaceWorkers = cfg.System.UserspaceDataplane.Workers
-				if cfg.System.UserspaceDataplane.RSSIndirectionDisabled {
-					rssEnabled = false
-				}
-				// Codex H1: scope D3 to only interfaces that
-				// userspace-dp actually binds AF_XDP sockets on.
-				rssAllowed = dpuserspace.UserspaceBoundLinuxInterfaces(cfg)
-				// #801 knobs.
-				claimHostTunables = cfg.System.UserspaceDataplane.ClaimHostTunables
-				governor = cfg.System.UserspaceDataplane.CPUGovernor
-				netdevBudget = cfg.System.UserspaceDataplane.NetdevBudget
-				coalesceExplicit = cfg.System.UserspaceDataplane.CoalescenceAdaptiveExplicit
-				// coalesceEnable stays false by default — the Step-0
-				// finding is "adaptive=on causes pp99 latency jitter",
-				// so default-off is what the issue asks for. An
-				// explicit `adaptive enable` inverts this.
-				if coalesceExplicit &&
-					!cfg.System.UserspaceDataplane.CoalescenceAdaptiveDisabled {
-					coalesceEnable = true
-				}
-				coalesceRX = cfg.System.UserspaceDataplane.CoalescenceRXUsecs
-				coalesceTX = cfg.System.UserspaceDataplane.CoalescenceTXUsecs
-			}
-		}
-		if err := enumerateAndRenameInterfaces(nodeID, clusterMode, userspaceWorkers, rssEnabled, rssAllowed); err != nil {
-			slog.Warn("interface naming failed", "err", err)
-		}
-		// #801: host tunables + coalescence. Runs after the interface
-		// rename but still before the dataplane is loaded — matches
-		// the D3 "before any AF_XDP bind" invariant. Best-effort: any
-		// failure logs and continues.
-		//
-		// B1 opt-in gate: host-scope knobs (governor + netdev_budget +
-		// adaptive-rx/tx flip) only apply when `claim-host-tunables
-		// true` is set. This keeps xpfd from stepping on shared hosts
-		// silently. D3 and per-iface rx-usecs/tx-usecs continue to run
-		// as before — both are interface-scoped.
-		d.applyStep0Tunables(userspaceDP, claimHostTunables, governor, netdevBudget,
-			coalesceExplicit, coalesceEnable, coalesceRX, coalesceTX, rssAllowed)
-	}
-
-	// Initialize routing, FRR, and IPsec managers
-	if !d.opts.NoDataplane {
-		rm, err := routing.New()
-		if err != nil {
-			slog.Warn("failed to create routing manager", "err", err)
-		} else {
-			d.routing = rm
-		}
-		d.frr = frr.New()
-		d.ipsec = ipsec.New()
-		d.ra = ra.New()
-		d.networkd = networkd.New()
-		d.dhcpServer = dhcpserver.New()
-	}
-
-	// Initialize cluster manager if configured (heartbeat/sync started after applyConfig).
-	if cfg := d.store.ActiveConfig(); cfg != nil && cfg.Chassis.Cluster != nil {
-		cc := cfg.Chassis.Cluster
-		d.cluster = cluster.NewManager(cc.NodeID, cc.ClusterID)
-		d.cluster.SetSoftwareVersion(d.opts.Version)
-		d.cluster.UpdateConfig(cc)
-		d.cluster.Start(ctx)
-		// Wire event-drop callback: on dropped cluster events, trigger
-		// immediate reconciliation so the safety net doesn't wait 2s.
-		d.cluster.SetOnEventDrop(d.triggerReconcile)
-		slog.Info("cluster manager initialized",
-			"node", cc.NodeID, "cluster", cc.ClusterID)
-
-		// Watch cluster events for state transitions (primary/secondary).
-		go d.watchClusterEvents(ctx)
-	}
-
-	// Enable IP forwarding — required for the firewall to route packets.
-	if !d.opts.NoDataplane {
-		enableForwarding()
-	}
-
-	// Create VRRP manager eagerly — must exist before applyConfig runs.
-	d.vrrpMgr = vrrp.NewManager()
-	// Wire event-drop callback: on dropped VRRP events, trigger
-	// immediate reconciliation.
-	d.vrrpMgr.SetOnEventDrop(d.triggerReconcile)
-	if err := d.vrrpMgr.Start(context.Background()); err != nil {
-		slog.Warn("failed to start VRRP manager", "err", err)
-	}
-	// On fresh cluster daemon start, suppress VRRP preemption until session
-	// bulk sync completes (or timeout) to avoid preempt-before-sync outages.
-	// Only applies when VRRP is enabled — otherwise no RETH VRRP instances.
-	if cfg := d.store.ActiveConfig(); cfg != nil && cfg.Chassis.Cluster != nil {
-		cc := cfg.Chassis.Cluster
-		if cc.FabricInterface != "" && cc.FabricPeerAddress != "" && !cc.NoRethVRRP && !cc.PrivateRGElection {
-			d.vrrpMgr.SetSyncHold(30 * time.Second)
-		}
-		// Private-rg-election mode: gate RG promotion on session sync
-		// readiness with a 30s timeout fallback (mirrors VRRP sync-hold).
-		// Without this, standalone nodes or nodes with permanently-down
-		// peers would never become primary.
-		if cc.PrivateRGElection && cc.FabricInterface != "" && cc.FabricPeerAddress != "" {
-			d.armSyncReadyTimer()
-		}
-	}
-
-	// Create dataplane backend (unless in config-only mode)
-	if !d.opts.NoDataplane {
-		dpType := ""
-		if cfg := d.store.ActiveConfig(); cfg != nil {
-			dpType = cfg.System.DataplaneType
-		}
-		dp, err := dataplane.NewDataPlane(dpType)
-		if err != nil {
-			slog.Error("failed to create dataplane", "type", dpType, "err", err)
-			return fmt.Errorf("create dataplane: %w", err)
-		}
-		d.dp = dp
-		if err := d.dp.Load(); err != nil {
-			slog.Warn("failed to load dataplane programs, running in config-only mode",
-				"err", err)
-			d.dp = nil
-		} else {
-			d.dp.SeedNATPortCounters()
-			nodeID := 0
-			if cfg := d.store.ActiveConfig(); cfg != nil && cfg.Chassis.Cluster != nil {
-				nodeID = cfg.Chassis.Cluster.NodeID
-			}
-			d.dp.SeedSessionIDCounter(nodeID)
-		}
-		// Apply current config — needed even in config-only mode so that
-		// VRFs, interfaces, and routing are configured before cluster comms.
-		if cfg := d.store.ActiveConfig(); cfg != nil {
-			slog.Info("applying active configuration")
-			d.applyConfig(cfg)
-		}
-	}
-
-	// Remove stale blackhole routes from previous daemon runs before
-	// cluster comms start (which may inject new ones).
-	if d.cluster != nil {
-		d.reconcileBlackholeRoutes()
-	}
-
-	// Start cluster heartbeat + sync after applyConfig (needs VRF to exist).
-	if d.cluster != nil {
-		d.startClusterComms(ctx)
-	}
-
-	// Handle signals for clean shutdown.
-	// In interactive mode, only SIGTERM triggers shutdown — SIGINT is handled
-	// by the CLI for command cancellation (Ctrl-C).
-	// In daemon mode, both SIGTERM and SIGINT trigger shutdown.
-	var stop context.CancelFunc
-	if isInteractive() {
-		ctx, stop = signal.NotifyContext(ctx, syscall.SIGTERM)
-	} else {
-		ctx, stop = signal.NotifyContext(ctx, syscall.SIGTERM, syscall.SIGINT)
-	}
-	defer stop()
-
-	// Create event buffer (shared between event reader and CLI)
-	eventBuf := logging.NewEventBuffer(1000)
-
-	// WaitGroup for coordinated shutdown of background goroutines
-	var wg sync.WaitGroup
-
-	// NOTE: session sync dp wiring + sweep start moved into startClusterComms
-	// goroutine to avoid race: d.sessionSync is created asynchronously.
-
-	// Start background services if dataplane is loaded
-	var er *logging.EventReader
-	if d.dp != nil {
-		// Start FIB sync (DPDK: background route populator; eBPF: no-op)
-		d.dp.StartFIBSync(ctx)
-
-		gc := conntrack.NewGC(d.dp, 10*time.Second)
-		d.gc = gc
-
-		// When the userspace dataplane is active, skip BPF session map
-		// GC entirely — sessions are managed in user-space. Without
-		// this, BatchLookup burns ~19% CPU scanning maps not used for
-		// forwarding decisions.
-		//
-		// The helper still mirrors sessions to BPF conntrack for display
-		// and periodically refreshes last_seen (~10s) so IterateSessions
-		// callers see accurate idle times.  See #333.
-		if _, ok := d.dp.(userspaceSessionDeltaDrainer); ok {
-			gc.SkipSweep = func() bool { return true }
-		}
-
-		// In cluster mode, GC should only expire sessions when this node
-		// is primary.  The peer primary ages sessions and syncs deletes.
-		if d.cluster != nil {
-			gc.IsLocalPrimary = d.cluster.IsLocalPrimaryAny
-		}
-
-		// Wire GC delete callbacks for incremental session sync.
-		// Deletes are synced if this node is primary for any RG — the peer
-		// ignores deletes for sessions it doesn't have.
-		gc.OnDeleteV4 = func(key dataplane.SessionKey) {
-			// Always sync deletes. Dropping deletes leaves stale sessions
-			// on the peer indefinitely.
-			if d.cluster != nil && d.cluster.IsLocalPrimaryAny() && d.sessionSync != nil {
-				d.sessionSync.QueueDeleteV4(key)
-			}
-		}
-		gc.OnDeleteV6 = func(key dataplane.SessionKeyV6) {
-			if d.cluster != nil && d.cluster.IsLocalPrimaryAny() && d.sessionSync != nil {
-				d.sessionSync.QueueDeleteV6(key)
-			}
-		}
-
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			gc.Run(ctx)
-		}()
-
-		evSrc, evErr := d.dp.NewEventSource()
-		if evErr != nil {
-			slog.Warn("failed to create event source", "err", evErr)
-		}
-		if evSrc != nil {
-			er = logging.NewEventReader(evSrc, eventBuf)
-			d.eventReader = er
-			wg.Add(1)
-			go func() {
-				defer wg.Done()
-				er.Run(ctx)
-			}()
-
-			// Wire ring buffer callback for near-real-time session sync.
-			if d.sessionSync != nil {
-				er.AddCallback(func(rec logging.EventRecord, raw []byte) {
-					if rec.Type != "SESSION_OPEN" {
-						return
-					}
-					if d.cluster == nil || !d.cluster.IsLocalPrimaryAny() {
-						return
-					}
-					if !d.sessionSync.IsConnected() {
-						return
-					}
-					if len(raw) < 56 {
-						return
-					}
-					proto := raw[53]
-					af := raw[55]
-					if af == dataplane.AFInet6 {
-						var key dataplane.SessionKeyV6
-						copy(key.SrcIP[:], raw[8:24])
-						copy(key.DstIP[:], raw[24:40])
-						key.SrcPort = binary.BigEndian.Uint16(raw[40:42])
-						key.DstPort = binary.BigEndian.Uint16(raw[42:44])
-						key.Protocol = proto
-						if val, err := d.dp.GetSessionV6(key); err == nil && val.IsReverse == 0 {
-							if d.sessionSync.ShouldSyncZone(val.IngressZone) {
-								d.sessionSync.QueueSessionV6(key, val)
-							}
-						}
-					} else {
-						var key dataplane.SessionKey
-						copy(key.SrcIP[:], raw[8:12])
-						copy(key.DstIP[:], raw[24:28])
-						key.SrcPort = binary.BigEndian.Uint16(raw[40:42])
-						key.DstPort = binary.BigEndian.Uint16(raw[42:44])
-						key.Protocol = proto
-						if val, err := d.dp.GetSessionV4(key); err == nil && val.IsReverse == 0 {
-							if d.sessionSync.ShouldSyncZone(val.IngressZone) {
-								d.sessionSync.QueueSessionV4(key, val)
-							}
-						}
-					}
-				})
-			}
-
-			// Set up syslog clients from active config
-			if cfg := d.store.ActiveConfig(); cfg != nil {
-				d.applySyslogConfig(er, cfg)
-			}
-
-			// Start NetFlow exporter if configured
-			if cfg := d.store.ActiveConfig(); cfg != nil {
-				d.startFlowExporter(ctx, cfg, er)
-			}
-
-			// Start IPFIX exporter if configured
-			if cfg := d.store.ActiveConfig(); cfg != nil {
-				d.startIPFIXExporter(ctx, cfg, er)
-			}
-
-			// Set up flow traceoptions if configured
-			if cfg := d.store.ActiveConfig(); cfg != nil {
-				d.applyFlowTrace(cfg, er)
-			}
-		}
-	}
-
-	// Start DHCP clients for interfaces configured with dhcp/dhcpv6.
-	// This must happen after BPF load + config compile so HOST_INBOUND_DHCP
-	// flags are active before DHCP packets start flowing.
-	if !d.opts.NoDataplane {
-		if cfg := d.store.ActiveConfig(); cfg != nil {
-			d.startDHCPClients(ctx, cfg)
-		}
-	}
-
-	// Start dynamic address feeds if configured.
-	if cfg := d.store.ActiveConfig(); cfg != nil && len(cfg.Security.DynamicAddress.FeedServers) > 0 {
-		d.feeds = feeds.New(func() {
-			slog.Info("dynamic-address feed updated, recompiling dataplane")
-			if activeCfg := d.store.ActiveConfig(); activeCfg != nil {
-				d.applyConfig(activeCfg)
-			}
-		})
-		d.feeds.Apply(ctx, &cfg.Security.DynamicAddress)
-	}
-
-	// Start RPM probes if configured.
-	if cfg := d.store.ActiveConfig(); cfg != nil && cfg.Services.RPM != nil && len(cfg.Services.RPM.Probes) > 0 {
-		d.rpm = rpm.New()
-		d.rpm.Apply(ctx, cfg.Services.RPM)
-	}
-
-	// Start LLDP if configured.
-	if cfg := d.store.ActiveConfig(); cfg != nil && cfg.Protocols.LLDP != nil && !cfg.Protocols.LLDP.Disable && len(cfg.Protocols.LLDP.Interfaces) > 0 {
-		d.lldpMgr = lldp.New()
-		var lldpIfaces []lldp.LLDPInterface
-		for _, iface := range cfg.Protocols.LLDP.Interfaces {
-			lldpIfaces = append(lldpIfaces, lldp.LLDPInterface{
-				Name:    iface.Name,
-				Disable: iface.Disable,
-			})
-		}
-		d.lldpMgr.Apply(ctx, &lldp.LLDPConfig{
-			Interfaces:     lldpIfaces,
-			Interval:       cfg.Protocols.LLDP.Interval,
-			HoldMultiplier: cfg.Protocols.LLDP.HoldMultiplier,
-			SystemName:     cfg.System.HostName,
-		})
-	}
-
-	// Start event-options engine if configured.
-	if cfg := d.store.ActiveConfig(); cfg != nil && len(cfg.EventOptions) > 0 {
-		// #846: route through commitAndApply so the engine's commit
-		// serializes with HTTP/gRPC commits under d.applySem.
-		// Event-options changes don't sync to peer (the engine fires
-		// independently on each node based on local RPM events).
-		d.eventEngine = eventengine.New(d.store, func(ctx context.Context, comment string) (*config.Config, error) {
-			return d.commitAndApply(ctx, comment, false)
-		})
-		d.eventEngine.Apply(cfg.EventOptions)
-		if d.rpm != nil {
-			d.rpm.SetEventCallback(d.eventEngine.HandleEvent)
-		}
-		slog.Info("event-options engine started", "policies", len(cfg.EventOptions))
-	}
-
-	// Start DHCP relay if configured.
-	if cfg := d.store.ActiveConfig(); cfg != nil && cfg.ForwardingOptions.DHCPRelay != nil {
-		d.dhcpRelay = dhcprelay.NewManager()
-		d.dhcpRelay.Apply(ctx, cfg.ForwardingOptions.DHCPRelay)
-	}
-
-	// Port mirroring
-	if cfg := d.store.ActiveConfig(); cfg != nil && cfg.ForwardingOptions.PortMirroring != nil {
-		for name, inst := range cfg.ForwardingOptions.PortMirroring.Instances {
-			slog.Info("Port mirroring configured", "instance", name, "input", inst.Input, "output", inst.Output)
-		}
-	}
-
-	// Start SNMP agent if configured (unless system processes snmp disable).
-	if cfg := d.store.ActiveConfig(); cfg != nil && cfg.System.SNMP != nil && (len(cfg.System.SNMP.Communities) > 0 || len(cfg.System.SNMP.V3Users) > 0) && !isProcessDisabled(cfg, "snmpd") {
-		d.snmpAgent = snmp.NewAgent(cfg.System.SNMP)
-		d.snmpAgent.SetIfDataFn(func() []snmp.IfData {
-			links, err := netlink.LinkList()
-			if err != nil {
-				return nil
-			}
-			var result []snmp.IfData
-			for _, link := range links {
-				attrs := link.Attrs()
-				if attrs.Name == "lo" {
-					continue
-				}
-				ifType := 6 // ethernetCsmacd
-				switch link.Type() {
-				case "vrf":
-					ifType = 53 // propVirtual
-				case "gre", "ip6tnl", "xfrm":
-					ifType = 131 // tunnel
-				case "veth":
-					ifType = 53
-				}
-				admin := 2 // down
-				if attrs.Flags&net.FlagUp != 0 {
-					admin = 1
-				}
-				oper := 2 // down
-				if attrs.OperState == netlink.OperUp || attrs.OperState == netlink.OperUnknown {
-					oper = 1
-				}
-				speed := uint32(0)
-				if attrs.TxQLen > 0 {
-					speed = 1000000000 // default 1Gbps
-				}
-				var stats *netlink.LinkStatistics
-				if attrs.Statistics != nil {
-					stats = attrs.Statistics
-				}
-				entry := snmp.IfData{
-					IfIndex:     attrs.Index,
-					IfDescr:     attrs.Name,
-					IfType:      ifType,
-					IfMtu:       attrs.MTU,
-					IfSpeed:     speed,
-					AdminStatus: admin,
-					OperStatus:  oper,
-					IfName:      attrs.Name,
-					IfHighSpeed: speed / 1_000_000, // bps -> Mbps
-				}
-				if stats != nil {
-					entry.InOctets = uint32(stats.RxBytes)
-					entry.OutOctets = uint32(stats.TxBytes)
-					entry.HCInOctets = stats.RxBytes
-					entry.HCInUcastPkts = stats.RxPackets
-					entry.HCOutOctets = stats.TxBytes
-					entry.HCOutUcastPkts = stats.TxPackets
-					entry.InMulticastPkts = uint32(stats.Multicast)
-				}
-				result = append(result, entry)
-			}
-			return result
-		})
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			d.snmpAgent.Start(ctx)
-		}()
-
-		// Start link state monitor for SNMP traps.
-		if len(cfg.System.SNMP.TrapGroups) > 0 {
-			wg.Add(1)
-			go func() {
-				defer wg.Done()
-				d.monitorLinkState(ctx)
-			}()
-		}
-	}
-
-	// Start policy scheduler if configured.
-	if cfg := d.store.ActiveConfig(); cfg != nil && len(cfg.Schedulers) > 0 && d.dp != nil {
-		d.scheduler = scheduler.New(cfg.Schedulers, func(activeState map[string]bool) {
-			slog.Info("scheduler state changed, updating policy rules")
-			if activeCfg := d.store.ActiveConfig(); activeCfg != nil {
-				d.dp.UpdatePolicyScheduleState(activeCfg, activeState)
-			}
-		})
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			d.scheduler.Run(ctx)
-		}()
-	}
-
-	// Start periodic neighbor resolution to keep ARP entries warm for
-	// known forwarding targets (DNAT pools, gateways, address-book hosts).
-	// Without this, bpf_fib_lookup returns NO_NEIGH when ARP expires,
-	// causing cold-start delays or connection failures for return traffic.
-	if !d.opts.NoDataplane {
-		if cfg := d.store.ActiveConfig(); cfg != nil {
-			wg.Add(1)
-			go func() {
-				defer wg.Done()
-				d.runPeriodicNeighborResolution(ctx)
-			}()
-		}
-	}
-
-	// Start VRRP event watcher (manager was created earlier, before applyConfig).
-	// Uses context.Background() — the watcher must outlive daemon ctx cancel
-	// so it can process VRRP BACKUP events during shutdown (rg_active cleanup).
-	// The watcher exits when eventCh is closed by vrrpMgr.Stop().
-	go d.watchVRRPEvents(context.Background())
-
-	// Start reconciliation loop — periodic safety net that corrects
-	// rg_active and blackhole route drift from dropped events.
-	if d.cluster != nil {
-		go d.reconcileRGStateLoop(ctx)
-	}
-
-	// Start HTTP API server if configured.
-	if d.opts.APIAddr != "" {
-		apiCfg := api.Config{
-			Addr:     d.opts.APIAddr,
-			Store:    d.store,
-			DP:       d.dp,
-			EventBuf: eventBuf,
-			GC:       d.gc,
-			Routing:  d.routing,
-			FRR:      d.frr,
-			IPsec:    d.ipsec,
-			DHCP:     d.dhcp,
-			VRRPMgr:  d.vrrpMgr,
-			// HTTP commits don't sync to peer (preserves prior
-			// behavior; see #846 for follow-up).
-			CommitFn: func(ctx context.Context, comment string) (*config.Config, error) {
-				return d.commitAndApply(ctx, comment, false)
-			},
-			CommitConfirmedFn: func(ctx context.Context, minutes int) (*config.Config, error) {
-				return d.commitConfirmedAndApply(ctx, minutes, false)
-			},
-			// #758: surface compile state so /health returns 503
-			// when the dataplane has never compiled successfully.
-			CompileHealthFn: func() api.CompileHealthSnapshot {
-				h := d.CompileHealthSnapshot()
-				return api.CompileHealthSnapshot{
-					EverSucceeded:    h.EverSucceeded,
-					FailureCount:     h.FailureCount,
-					LastError:        h.LastError,
-					LastErrorUnixSec: h.LastErrorUnixSec,
-				}
-			},
-		}
-		// Resolve interface bindings from web-management config
-		if cfg := d.store.ActiveConfig(); cfg != nil && cfg.System.Services != nil &&
-			cfg.System.Services.WebManagement != nil {
-			wm := cfg.System.Services.WebManagement
-			// Bind HTTP to configured interface
-			if wm.HTTPInterface != "" {
-				bindIP := resolveInterfaceAddr(wm.HTTPInterface, "127.0.0.1")
-				apiCfg.Addr = bindIP + ":8080"
-				slog.Info("HTTP API bound to interface", "interface", wm.HTTPInterface, "addr", apiCfg.Addr)
-			}
-			// Enable HTTPS if configured
-			if wm.HTTPS {
-				httpsBindIP := "127.0.0.1"
-				if wm.HTTPSInterface != "" {
-					httpsBindIP = resolveInterfaceAddr(wm.HTTPSInterface, "127.0.0.1")
-					slog.Info("HTTPS API bound to interface", "interface", wm.HTTPSInterface, "addr", httpsBindIP+":8443")
-				}
-				apiCfg.TLS = true
-				apiCfg.HTTPSAddr = httpsBindIP + ":8443"
-			}
-			// API authentication
-			if wm.APIAuth != nil && (len(wm.APIAuth.Users) > 0 || len(wm.APIAuth.APIKeys) > 0) {
-				authCfg := &api.AuthConfig{
-					Users:   make(map[string]string),
-					APIKeys: make(map[string]bool),
-				}
-				for _, u := range wm.APIAuth.Users {
-					authCfg.Users[u.Username] = u.Password
-				}
-				for _, k := range wm.APIAuth.APIKeys {
-					authCfg.APIKeys[k] = true
-				}
-				apiCfg.Auth = authCfg
-				slog.Info("HTTP API authentication enabled", "users", len(wm.APIAuth.Users), "api_keys", len(wm.APIAuth.APIKeys))
-			}
-		}
-		srv := api.NewServer(apiCfg)
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			if err := srv.Run(ctx); err != nil {
-				slog.Error("API server error", "err", err)
-			}
-		}()
-		slog.Info("HTTP API server started", "addr", d.opts.APIAddr)
-	}
-
-	// #881: forwarding-daemon CPU sampler (5s/1m/5m windows for
-	// `show chassis forwarding`).  Shared between the gRPC server
-	// and the local CLI; both paths call Snapshot() at query time.
-	// Started here so the ring is populated before the first CLI.
-	fwdSampler := fwdstatus.NewSampler(d.dp, fwdstatus.OSProcReader{})
-	fwdSampler.Start(ctx)
-
-	// Start gRPC API server.
-	{
-		grpcSrv := grpcapi.NewServer(d.opts.GRPCAddr, grpcapi.Config{
-			Store:      d.store,
-			DP:         d.dp,
-			EventBuf:   eventBuf,
-			GC:         d.gc,
-			Routing:    d.routing,
-			FRR:        d.frr,
-			IPsec:      d.ipsec,
-			Cluster:    d.cluster,
-			DHCP:       d.dhcp,
-			DHCPServer: d.dhcpServer,
-			RPMResultsFn: func() []*rpm.ProbeResult {
-				if d.rpm != nil {
-					return d.rpm.Results()
-				}
-				return nil
-			},
-			FeedsFn: func() map[string]feeds.FeedInfo {
-				if d.feeds != nil {
-					return d.feeds.AllFeeds()
-				}
-				return nil
-			},
-			LLDPNeighborsFn: func() []*lldp.Neighbor {
-				if d.lldpMgr != nil {
-					return d.lldpMgr.Neighbors()
-				}
-				return nil
-			},
-			// gRPC commits sync to cluster peer atomically inside
-			// the apply lock so the peer can never observe an apply
-			// that hasn't yet been propagated.
-			CommitFn: func(ctx context.Context, comment string) (*config.Config, error) {
-				return d.commitAndApply(ctx, comment, true)
-			},
-			CommitConfirmedFn: func(ctx context.Context, minutes int) (*config.Config, error) {
-				return d.commitConfirmedAndApply(ctx, minutes, true)
-			},
-			VRRPMgr: d.vrrpMgr,
-			RAMgr:   d.ra,
-			Version: d.opts.Version,
-			FabricPeerAddrFn: func() []string {
-				var addrs []string
-				if d.syncPeerAddr != "" {
-					addrs = append(addrs, d.syncPeerAddr)
-				} else {
-					d.fabricMu.RLock()
-					if d.fabricPeerIP != nil {
-						addrs = append(addrs, d.fabricPeerIP.String())
-					}
-					d.fabricMu.RUnlock()
-				}
-				if d.syncPeerAddr1 != "" {
-					addrs = append(addrs, d.syncPeerAddr1)
-				} else {
-					d.fabricMu.RLock()
-					if d.fabricPeerIP1 != nil {
-						addrs = append(addrs, d.fabricPeerIP1.String())
-					}
-					d.fabricMu.RUnlock()
-				}
-				return addrs
-			},
-			FabricVRFDevice: func() string {
-				if c := d.store.ActiveConfig(); c != nil && c.Chassis.Cluster != nil {
-					cc := c.Chassis.Cluster
-					if cc.ControlInterface != "" || cc.FabricInterface != "" {
-						return "vrf-mgmt"
-					}
-				}
-				return ""
-			}(),
-			FwdSampler: fwdSampler,
-		})
-		d.grpcSrv = grpcSrv
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			if err := grpcSrv.Run(ctx); err != nil {
-				slog.Error("gRPC server error", "err", err)
-			}
-		}()
-		slog.Info("gRPC API server started", "addr", d.opts.GRPCAddr)
-	}
-
-	// Start interactive CLI or block in daemon mode
-	var runErr error
-	if isInteractive() {
-		shell := cli.New(d.store, d.dp, eventBuf, er, d.routing, d.frr, d.ipsec, d.dhcp, d.dhcpRelay, d.cluster)
-		shell.SetVersion(d.opts.Version)
-		shell.SetForwardingSampler(fwdSampler)
-		// #797 H2 / #846: route in-process CLI commits through the
-		// daemon's atomic commit+apply so they serialize against
-		// HTTP/gRPC/event-engine commits under d.applySem.
-		// applyConfigFn stays wired for non-commit paths (rollback,
-		// confirm) that still need the full reconcile.
-		shell.SetApplyConfigFn(d.applyConfig)
-		shell.SetCommitFns(
-			func(ctx context.Context, comment string) (*config.Config, error) {
-				// In-process CLI commits don't sync to peer (the
-				// CLI is local; preserves prior per-transport
-				// behavior, same as HTTP).
-				return d.commitAndApply(ctx, comment, false)
-			},
-			func(ctx context.Context, minutes int) (*config.Config, error) {
-				return d.commitConfirmedAndApply(ctx, minutes, false)
-			},
-		)
-		shell.SetRPMResultsFn(func() []*rpm.ProbeResult {
-			if d.rpm != nil {
-				return d.rpm.Results()
-			}
-			return nil
-		})
-		shell.SetFeedsFn(func() map[string]feeds.FeedInfo {
-			if d.feeds != nil {
-				return d.feeds.AllFeeds()
-			}
-			return nil
-		})
-		shell.SetLLDPNeighborsFn(func() []*lldp.Neighbor {
-			if d.lldpMgr != nil {
-				return d.lldpMgr.Neighbors()
-			}
-			return nil
-		})
-		shell.SetVRRPManager(d.vrrpMgr)
-		shell.SetFabricPeer(func() []string {
-			var addrs []string
-			if d.syncPeerAddr != "" {
-				addrs = append(addrs, d.syncPeerAddr)
-			} else {
-				d.fabricMu.RLock()
-				if d.fabricPeerIP != nil {
-					addrs = append(addrs, d.fabricPeerIP.String())
-				}
-				d.fabricMu.RUnlock()
-			}
-			if d.syncPeerAddr1 != "" {
-				addrs = append(addrs, d.syncPeerAddr1)
-			} else {
-				d.fabricMu.RLock()
-				if d.fabricPeerIP1 != nil {
-					addrs = append(addrs, d.fabricPeerIP1.String())
-				}
-				d.fabricMu.RUnlock()
-			}
-			return addrs
-		}, func() string {
-			if c := d.store.ActiveConfig(); c != nil && c.Chassis.Cluster != nil {
-				cc := c.Chassis.Cluster
-				if cc.ControlInterface != "" || cc.FabricInterface != "" {
-					return "vrf-mgmt"
-				}
-			}
-			return ""
-		}())
-
-		// Set RBAC login class from config (default to super-user if user not found)
-		if cfg := d.store.ActiveConfig(); cfg != nil && cfg.System.Login != nil {
-			osUser := os.Getenv("USER")
-			found := false
-			for _, u := range cfg.System.Login.Users {
-				if u.Name == osUser {
-					shell.SetUserClass(u.Class)
-					found = true
-					break
-				}
-			}
-			if !found {
-				shell.SetUserClass("super-user")
-			}
-		}
-
-		// Run CLI in a goroutine so we can still handle signals
-		errCh := make(chan error, 1)
-		go func() {
-			errCh <- shell.Run()
-		}()
-
-		select {
-		case err := <-errCh:
-			if err != nil {
-				runErr = fmt.Errorf("CLI: %w", err)
-			}
-		case <-ctx.Done():
-			slog.Info("signal received, shutting down")
-		}
-	} else {
-		slog.Info("daemon mode (non-interactive), waiting for signals")
-		<-ctx.Done()
-		slog.Info("signal received, shutting down")
-	}
-
-	// Cancel context to stop background goroutines, then wait for them.
-	stop()
-	wg.Wait()
-
-	// Clean up flow exporters.
-	d.stopFlowExporter()
-	d.stopIPFIXExporter()
-
-	// Clean up dynamic address feeds.
-	if d.feeds != nil {
-		d.feeds.StopAll()
-	}
-
-	// Clean up RPM probes.
-	if d.rpm != nil {
-		d.rpm.StopAll()
-	}
-
-	// Clean up LLDP.
-	if d.lldpMgr != nil {
-		d.lldpMgr.Stop()
-	}
-
-	// Determine shutdown mode early so we can clear rg_active BEFORE
-	// stopping subsystems (VRRP, sync) that may hang.
-	cfg := d.store.ActiveConfig()
-	haMode := cfg != nil && cfg.Chassis.Cluster != nil
-	hitless := !haMode // standalone = hitless by default
-	if haMode && cfg.Chassis.Cluster.HitlessRestart {
-		hitless = true // operator explicitly opted in
-	}
-
-	// In HA fail-closed mode, clear rg_active and watchdog immediately so
-	// BPF stops forwarding traffic even if subsequent cleanup steps hang.
-	if !hitless && d.dp != nil && cfg.Chassis.Cluster != nil {
-		slog.Info("HA shutdown: clearing rg_active for all RGs")
-		for _, rg := range cfg.Chassis.Cluster.RedundancyGroups {
-			if err := d.dp.UpdateRGActive(rg.ID, false); err != nil {
-				slog.Warn("failed to clear rg_active on shutdown", "rg", rg.ID, "err", err)
-			}
-			if err := d.dp.UpdateHAWatchdog(rg.ID, 0); err != nil {
-				slog.Warn("failed to clear ha_watchdog on shutdown", "rg", rg.ID, "err", err)
-			}
-		}
-	}
-
-	// Withdraw RA senders (sends goodbye RAs with lifetime=0) before VRRP
-	// stop so hosts immediately stop using this node as a default router.
-	if d.ra != nil {
-		if err := d.ra.Withdraw(); err != nil {
-			slog.Warn("shutdown: failed to withdraw RA senders", "err", err)
-		}
-	}
-
-	// Direct-mode: remove VIPs before VRRP stop (VRRP won't manage them).
-	if d.isNoRethVRRP() && cfg.Chassis.Cluster != nil {
-		for _, rg := range cfg.Chassis.Cluster.RedundancyGroups {
-			d.directRemoveVIPs(rg.ID)
-		}
-	}
-
-	// Stop VRRP manager (removes VIPs, sends priority-0).
-	if d.vrrpMgr != nil {
-		d.vrrpMgr.Stop()
-	}
-
-	// Stop cluster monitor (heartbeats) immediately after VRRP priority-0.
-	// This ensures the peer's heartbeat timeout starts promptly instead of
-	// being delayed by the 5s sync Stop timeout below.
-	if d.cluster != nil {
-		d.cluster.Stop()
-	}
-
-	// Stop session sync (5s timeout to avoid blocking teardown).
-	if d.sessionSync != nil {
-		d.stopSyncReadyTimer()
-		d.sessionSync.Stop()
-	}
-
-	if d.dp != nil {
-		logFinalStats(d.dp)
-		if hitless {
-			// Hitless: close Go handles only — BPF programs keep running.
-			slog.Info("hitless shutdown: preserving BPF state")
-			d.dp.Close()
-		} else {
-			// Fail-closed: tear down all pinned BPF state.
-			slog.Info("HA shutdown: tearing down BPF state")
-			d.dp.Teardown()
-		}
-	}
-
-	// #801 B2: restore any host-scope tunables xpfd claimed to their
-	// pre-xpfd values. No-op if `claim-host-tunables` was never set.
-	// Runs on every shutdown (hitless + fail-closed) so stopping xpfd
-	// leaves the host as xpfd found it.
-	d.restoreStep0TunablesOnShutdown()
-
-	slog.Info("shutdown complete")
-	return runErr
-}
-
-// isInteractive returns true if stdin is a real terminal (not /dev/null or a pipe).
-// enableForwarding enables IPv4 and IPv6 forwarding via sysctl
-// and disables RA acceptance on all interfaces.
-// A firewall must forward packets between interfaces; without this,
-// the kernel drops all transit traffic. A firewall must not accept
-// RAs — it uses its own configured routes exclusively.
-func enableForwarding() {
-	sysctls := map[string]string{
-		"/proc/sys/net/ipv4/ip_forward":             "1",
-		"/proc/sys/net/ipv6/conf/all/forwarding":    "1",
-		"/proc/sys/net/ipv6/conf/all/accept_ra":     "0",
-		"/proc/sys/net/ipv6/conf/default/accept_ra": "0",
-		// l3mdev_accept: allow accepting TCP/UDP connections on management VRF
-		// interfaces from sockets not bound to the VRF (needed for SSH).
-		"/proc/sys/net/ipv4/tcp_l3mdev_accept": "1",
-		"/proc/sys/net/ipv4/udp_l3mdev_accept": "1",
-		// accept_local: allow packets with a source IP that is local to the
-		// machine on a different interface. Required when XDP SNAT rewrites
-		// src to a tunnel endpoint IP and XDP_PASS to kernel for routing —
-		// kernel would otherwise reject the packet as a martian.
-		"/proc/sys/net/ipv4/conf/all/accept_local": "1",
-	}
-	for path, val := range sysctls {
-		if err := os.WriteFile(path, []byte(val), 0644); err != nil {
-			slog.Warn("failed to set sysctl", "path", path, "err", err)
-		}
-	}
-	slog.Info("IP forwarding enabled, RA acceptance disabled")
-}
-
-func inferIPv6StaticNextHopInterfaces(cfg *config.Config) map[string]map[string]string {
-	type connectedPrefix struct {
-		net    *net.IPNet
-		ifName string
-		bits   int
-	}
-
-	var connected []connectedPrefix
-	connectedByLogical := make(map[string][]connectedPrefix)
-	ifNames := make([]string, 0, len(cfg.Interfaces.Interfaces))
-	for ifName := range cfg.Interfaces.Interfaces {
-		ifNames = append(ifNames, ifName)
-	}
-	sort.Strings(ifNames)
-	for _, ifName := range ifNames {
-		ifc := cfg.Interfaces.Interfaces[ifName]
-		base := config.LinuxIfName(ifName)
-		unitNums := make([]int, 0, len(ifc.Units))
-		for unitNum := range ifc.Units {
-			unitNums = append(unitNums, unitNum)
-		}
-		sort.Ints(unitNums)
-		for _, unitNum := range unitNums {
-			unit := ifc.Units[unitNum]
-			logical := base
-			if unitNum != 0 {
-				logical = fmt.Sprintf("%s.%d", base, unitNum)
-			}
-			for _, addr := range unit.Addresses {
-				ip, ipNet, err := net.ParseCIDR(addr)
-				if err != nil || ip == nil || ip.To4() != nil {
-					continue
-				}
-				bits, _ := ipNet.Mask.Size()
-				prefix := connectedPrefix{
-					net:    ipNet,
-					ifName: logical,
-					bits:   bits,
-				}
-				connected = append(connected, prefix)
-				connectedByLogical[logical] = append(connectedByLogical[logical], prefix)
-			}
-		}
-	}
-
-	resolve := func(candidates []connectedPrefix, addr string) string {
-		ip := net.ParseIP(addr)
-		if ip == nil || ip.To4() != nil {
-			return ""
-		}
-		bestIf := ""
-		bestBits := -1
-		for _, candidate := range candidates {
-			if !candidate.net.Contains(ip) {
-				continue
-			}
-			if candidate.bits > bestBits || (candidate.bits == bestBits && (bestIf == "" || candidate.ifName < bestIf)) {
-				bestIf = candidate.ifName
-				bestBits = candidate.bits
-			}
-		}
-		return bestIf
-	}
-
-	collectPrefixesForInterface := func(ifName string) []connectedPrefix {
-		normalized := config.LinuxIfName(ifName)
-		var prefixes []connectedPrefix
-		if entries, ok := connectedByLogical[normalized]; ok {
-			prefixes = append(prefixes, entries...)
-		}
-		if !strings.Contains(normalized, ".") {
-			prefixNames := make([]string, 0, len(connectedByLogical))
-			for logical := range connectedByLogical {
-				if strings.HasPrefix(logical, normalized+".") {
-					prefixNames = append(prefixNames, logical)
-				}
-			}
-			sort.Strings(prefixNames)
-			for _, logical := range prefixNames {
-				prefixes = append(prefixes, connectedByLogical[logical]...)
-			}
-		}
-		return prefixes
-	}
-
-	resolved := make(map[string]map[string]string)
-	connectedByVRF := map[string][]connectedPrefix{
-		"": append([]connectedPrefix(nil), connected...),
-	}
-	setResolved := func(vrfName, nextHop, ifName string) {
-		if ifName == "" {
-			return
-		}
-		vrfMap, ok := resolved[vrfName]
-		if !ok {
-			vrfMap = make(map[string]string)
-			resolved[vrfName] = vrfMap
-		}
-		if existing, ok := vrfMap[nextHop]; !ok || ifName < existing {
-			vrfMap[nextHop] = ifName
-		}
-	}
-	addRoutes := func(vrfName string, routes []*config.StaticRoute) {
-		candidates := connectedByVRF[vrfName]
-		for _, sr := range routes {
-			for _, nh := range sr.NextHops {
-				if nh.Interface != "" || nh.Address == "" || !strings.Contains(nh.Address, ":") {
-					continue
-				}
-				setResolved(vrfName, nh.Address, resolve(candidates, nh.Address))
-			}
-		}
-	}
-
-	claimedByVRF := make(map[string]struct{})
-	for _, ri := range cfg.RoutingInstances {
-		vrfName := "vrf-" + ri.Name
-		if ri.InstanceType == "forwarding" {
-			vrfName = ""
-		}
-		for _, ifName := range ri.Interfaces {
-			prefixes := collectPrefixesForInterface(ifName)
-			if len(prefixes) == 0 {
-				continue
-			}
-			connectedByVRF[vrfName] = append(connectedByVRF[vrfName], prefixes...)
-			if vrfName != "" {
-				normalized := config.LinuxIfName(ifName)
-				claimedByVRF[normalized] = struct{}{}
-			}
-		}
-	}
-	if len(claimedByVRF) > 0 {
-		filtered := connectedByVRF[""][:0]
-		for _, prefix := range connectedByVRF[""] {
-			base := prefix.ifName
-			if idx := strings.IndexByte(base, '.'); idx >= 0 {
-				base = base[:idx]
-			}
-			if _, claimed := claimedByVRF[prefix.ifName]; claimed {
-				continue
-			}
-			if _, claimed := claimedByVRF[base]; claimed {
-				continue
-			}
-			filtered = append(filtered, prefix)
-		}
-		connectedByVRF[""] = filtered
-	}
-
-	addRoutes("", cfg.RoutingOptions.StaticRoutes)
-	addRoutes("", cfg.RoutingOptions.Inet6StaticRoutes)
-	for _, ri := range cfg.RoutingInstances {
-		vrfName := "vrf-" + ri.Name
-		if ri.InstanceType == "forwarding" {
-			vrfName = ""
-		}
-		addRoutes(vrfName, ri.StaticRoutes)
-		addRoutes(vrfName, ri.Inet6StaticRoutes)
-	}
-	return resolved
-}
-
-// bootstrapFromFile reads the text Junos config file and imports it as the
-// initial active configuration. This is called on first start when the DB
-// has no active config yet.
-func (d *Daemon) bootstrapFromFile() error {
-	data, err := os.ReadFile(d.opts.ConfigFile)
-	if err != nil {
-		return fmt.Errorf("read config file: %w", err)
-	}
-
-	// Import into the store: enter config mode, load, commit.
-	// Commit() handles compilation (including ${node} variable expansion
-	// when nodeID is set on the store for cluster mode).
-	if err := d.store.EnterConfigure(); err != nil {
-		return fmt.Errorf("enter configure: %w", err)
-	}
-	if err := d.store.LoadOverride(string(data)); err != nil {
-		d.store.ExitConfigure()
-		return fmt.Errorf("load override: %w", err)
-	}
-	if _, err := d.store.Commit(); err != nil {
-		d.store.ExitConfigure()
-		return fmt.Errorf("commit: %w", err)
-	}
-	d.store.ExitConfigure()
-	slog.Info("configuration bootstrapped from file", "file", d.opts.ConfigFile)
-	return nil
-}
-
-// applyConfig applies a compiled config to the dataplane / kernel.
-// Wraps applyConfigLocked under the apply semaphore for non-context
-// callers (DHCP callbacks, config-poll, dynamic feeds, event engine,
-// in-process CLI commits, CLI auto-rollback, cluster sync recv).
-// Always succeeds in acquiring the lock because Background never
-// cancels.
-//
-// #846: HTTP/gRPC commit handlers go through commitAndApply /
-// commitConfirmedAndApply instead, which take the same semaphore
-// with a request-bound context so a slow lock holder surfaces 503
-// to the client rather than hanging the request.
-func (d *Daemon) applyConfig(cfg *config.Config) {
-	_ = d.applySem.Acquire(context.Background(), 1)
-	defer d.applySem.Release(1)
-	d.applyConfigLocked(cfg)
-}
-
-// commitAndApply atomically promotes the candidate config and
-// applies it. Holds applySem across configstore.Commit and
-// applyConfigLocked so two concurrent committers can't interleave
-// their commit→apply pairs (which would let kernel state lag store
-// state). Optionally syncs to the cluster peer inside the lock.
-//
-// If ctx is canceled before the semaphore is acquired, returns
-// ctx.Err() and NEITHER commit nor apply runs — no divergence. Once
-// the semaphore is held, commit and apply run to completion;
-// cancellation past that point is ignored (applyConfigLocked is not
-// safe to interrupt mid-stream — kernel route writes, FRR reload,
-// etc.).
-func (d *Daemon) commitAndApply(ctx context.Context, comment string, syncPeer bool) (*config.Config, error) {
-	if err := d.applySem.Acquire(ctx, 1); err != nil {
-		return nil, err
-	}
-	defer d.applySem.Release(1)
-
-	var compiled *config.Config
-	var err error
-	if comment != "" {
-		compiled, err = d.store.CommitWithDescription(comment)
-	} else {
-		compiled, err = d.store.Commit()
-	}
-	if err != nil {
-		return nil, err
-	}
-	d.applyConfigLocked(compiled)
-	if syncPeer {
-		d.syncConfigToPeer()
-	}
-	return compiled, nil
-}
-
-// syncAndApply is the cluster-sync-recv analogue of commitAndApply.
-// Holds applySem across configstore.SyncApply (peer-driven active
-// promotion) + applyConfigLocked, so a peer-sync can't interleave
-// between a local committer's Commit and applyConfig (which would
-// briefly leave store=peer-config but kernel=local-config).
-func (d *Daemon) syncAndApply(ctx context.Context, configText string, chassisPreserve func(*config.ConfigTree)) (*config.Config, error) {
-	if err := d.applySem.Acquire(ctx, 1); err != nil {
-		return nil, err
-	}
-	defer d.applySem.Release(1)
-
-	compiled, err := d.store.SyncApply(configText, chassisPreserve)
-	if err != nil {
-		return nil, err
-	}
-	if compiled != nil {
-		d.applyConfigLocked(compiled)
-	}
-	return compiled, nil
-}
-
-// commitConfirmedAndApply is the commit-confirmed analogue of
-// commitAndApply. Same atomicity guarantees.
-func (d *Daemon) commitConfirmedAndApply(ctx context.Context, minutes int, syncPeer bool) (*config.Config, error) {
-	if err := d.applySem.Acquire(ctx, 1); err != nil {
-		return nil, err
-	}
-	defer d.applySem.Release(1)
-
-	compiled, err := d.store.CommitConfirmed(minutes)
-	if err != nil {
-		return nil, err
-	}
-	d.applyConfigLocked(compiled)
-	if syncPeer {
-		d.syncConfigToPeer()
-	}
-	return compiled, nil
-}
-
-// applyConfigLocked runs the actual reconcile pipeline. MUST be
-// called with d.applySem held.
-func (d *Daemon) applyConfigLocked(cfg *config.Config) {
-	if d.applyBodyForTest != nil {
-		d.applyBodyForTest(cfg)
-		return
-	}
-	// Reset VIP warning suppression so new config gets fresh warnings.
-	d.vipWarnedIfaces = nil
-
-	// Log config validation warnings
-	for _, w := range cfg.Warnings {
-		slog.Warn("config validation", "warning", w)
-	}
-
-	// 0. Reconcile VRF devices (routing-instance VRFs + management VRF).
-	// ReconcileVRFs is idempotent: VRFs already present with the correct
-	// table ID are preserved (ifindex unchanged). Removed-from-config
-	// VRFs are deleted. #847: xpfd claims the entire `vrf-*` kernel
-	// namespace — orphan vrf-* devices not in desired and not in
-	// m.vrfs (e.g. left over from a routing-instance rename across
-	// a daemon restart) are also reaped. Operators MUST NOT
-	// pre-create vrf-<name> outside xpfd config.
-	//
-	// (The original docs/pr/844-vrf-idempotent/plan.md described an
-	// earlier design where external VRFs were left alone; the
-	// namespace-claim policy in this code supersedes that plan. See
-	// the godoc on routing.ReconcileVRFs for the current contract.)
-	const mgmtVRFName = "mgmt"
-	const mgmtTableID = 999
-	mgmtIfaces := make(map[string]bool)
-	for name := range cfg.Interfaces.Interfaces {
-		if strings.HasPrefix(name, "fxp") || strings.HasPrefix(name, "fab") || strings.HasPrefix(name, "em") {
-			mgmtIfaces[config.LinuxIfName(name)] = true
-		}
-	}
-
-	if d.routing != nil {
-		var desired []routing.VRFSpec
-		for _, ri := range cfg.RoutingInstances {
-			if ri.InstanceType == "forwarding" {
-				slog.Info("forwarding instance, skipping VRF creation",
-					"instance", ri.Name)
-				continue
-			}
-			desired = append(desired, routing.VRFSpec{
-				Name:    ri.Name,
-				TableID: ri.TableID,
-			})
-		}
-		if len(mgmtIfaces) > 0 {
-			desired = append(desired, routing.VRFSpec{
-				Name:    mgmtVRFName,
-				TableID: mgmtTableID,
-			})
-		}
-		if err := d.routing.ReconcileVRFs(desired); err != nil {
-			slog.Warn("failed to reconcile VRFs", "err", err)
-		}
-	}
-
-	// 0a. Bind routing-instance interfaces to their VRFs.
-	if d.routing != nil {
-		for _, ri := range cfg.RoutingInstances {
-			if ri.InstanceType == "forwarding" {
-				continue
-			}
-			for _, ifaceName := range ri.Interfaces {
-				// Convert Junos name (gr-0/0/0.0) to Linux name (gr-0-0-0).
-				// Strip ".0" unit suffix — unit 0 is the base interface.
-				linuxName := config.LinuxIfName(ifaceName)
-				if strings.HasSuffix(linuxName, ".0") {
-					linuxName = strings.TrimSuffix(linuxName, ".0")
-				}
-				if err := d.routing.BindInterfaceToVRF(linuxName, ri.Name); err != nil {
-					slog.Warn("failed to bind interface to VRF",
-						"interface", ifaceName, "linux", linuxName,
-						"instance", ri.Name, "err", err)
-				}
-			}
-		}
-	}
-
-	// 0b. Bind management interfaces (fxp*/fab*/em*) to vrf-mgmt, but
-	// only if ReconcileVRFs actually got vrf-mgmt into the managed set.
-	// If reconcile errored out before vrf-mgmt could be created,
-	// downstream code (applyMgmtVRFRoutes, HA sync) would otherwise
-	// run against a non-existent VRF.
-	d.mgmtVRFInterfaces = nil
-	if d.routing != nil && len(mgmtIfaces) > 0 && d.routing.IsManagedVRF(mgmtVRFName) {
-		d.mgmtVRFInterfaces = mgmtIfaces
-		for ifName := range mgmtIfaces {
-			if err := d.routing.BindInterfaceToVRF(ifName, mgmtVRFName); err != nil {
-				slog.Warn("failed to bind interface to management VRF",
-					"interface", ifName, "err", err)
-			}
-		}
-	}
-
-	// 0.6. Program default routes in the management VRF for DHCP leases.
-	d.applyMgmtVRFRoutes()
-
-	// 1. Create tunnel interfaces (interface-level + per-unit tunnels)
-	if d.routing != nil {
-		if err := d.routing.ApplyTunnels(collectAppliedTunnels(cfg)); err != nil {
-			slog.Warn("failed to apply tunnels", "err", err)
-		}
-	}
-
-	// 1.5. Create xfrmi interfaces for IPsec VPN tunnels.
-	// Must happen before BPF compilation so compileZones() can discover
-	// the xfrmi interfaces and map them to security zones.
-	// Always call ApplyXfrmi so stale xfrmi devices are removed when VPNs
-	// are deleted from config.
-	if d.routing != nil {
-		if err := d.routing.ApplyXfrmi(cfg.Security.IPsec.VPNs); err != nil {
-			slog.Warn("failed to apply xfrmi interfaces", "err", err)
-		}
-	}
-
-	// 1.7. Create bond (LAG) interfaces for fabric-options member-interfaces.
-	// Always call ApplyBonds (even with empty list) so stale bonds from
-	// previous configs get cleaned up via ClearBonds().
-	if d.routing != nil {
-		var bondIfaces []*config.InterfaceConfig
-		for _, ifc := range cfg.Interfaces.Interfaces {
-			if len(ifc.FabricMembers) > 0 {
-				bondIfaces = append(bondIfaces, ifc)
-			}
-		}
-		if err := d.routing.ApplyBonds(bondIfaces); err != nil {
-			slog.Warn("failed to apply bonds", "err", err)
-		}
-	}
-
-	// 1.8. Clean up legacy RETH bond devices from previous binary versions.
-	// VRRP now runs directly on physical member interfaces — no bonds needed.
-	if d.routing != nil {
-		d.routing.ClearRethInterfaces()
-	}
-
-	// 1.9. Create IPVLAN interfaces for fabric members (fab0, fab1).
-	// The physical member (ge-0-0-0) keeps its name; fab0 is IPVLAN L2
-	// on top for IP addressing. BPF attaches to the parent.
-	// Track which overlays are configured so stale ones can be cleaned up (#128).
-	//
-	// When the userspace dataplane is active, DEFER IPVLAN creation until
-	// after XSK binds complete. The kernel checks for upper devices (like
-	// IPVLAN) at XSK bind time — if an IPVLAN exists, zerocopy bind fails
-	// and falls back to copy mode (~3 Gbps). Deferring lets the fabric
-	// parent bind XSK in zerocopy first, then the IPVLAN is added for
-	// sync/heartbeat addressing.
-	activeFabricOverlays := make(map[string]bool)
-	type deferredIPVLAN struct {
-		parent string
-		name   string
-		addrs  []string
-	}
-	var deferredOverlays []deferredIPVLAN
-	_, isUserspaceDP := d.dp.(*dpuserspace.Manager)
-	for ifName, ifCfg := range cfg.Interfaces.Interfaces {
-		if ifCfg.LocalFabricMember == "" || !strings.HasPrefix(ifName, "fab") {
-			continue
-		}
-		parentLinux := config.LinuxIfName(ifCfg.LocalFabricMember)
-		fabLinux := config.LinuxIfName(ifName)
-		activeFabricOverlays[fabLinux] = true
-		var addrs []string
-		if unit, ok := ifCfg.Units[0]; ok {
-			addrs = unit.Addresses
-		}
-		// When userspace DP is active, remove any existing IPVLAN and
-		// defer recreation until after XSK binds in zerocopy. The kernel
-		// checks for upper devices at bind time — IPVLAN blocks zerocopy.
-		// On subsequent applyConfig calls (config change), the IPVLAN
-		// already exists from the OnXSKBound callback and XSK is already
-		// bound, so the xskBoundNotified guard prevents re-deletion.
-		if isUserspaceDP {
-			if um, ok := d.dp.(*dpuserspace.Manager); ok && !um.XSKBoundNotified() {
-				// First applyConfig — remove stale IPVLAN so XSK can zerocopy.
-				if link, err := netlink.LinkByName(fabLinux); err == nil {
-					netlink.LinkDel(link)
-					slog.Info("removed fabric IPVLAN for deferred zerocopy XSK bind",
-						"name", fabLinux)
-				}
-				deferredOverlays = append(deferredOverlays, deferredIPVLAN{
-					parent: parentLinux, name: fabLinux, addrs: addrs,
-				})
-				slog.Info("deferring fabric IPVLAN creation until XSK binds complete",
-					"parent", parentLinux, "name", fabLinux)
-				// continue // DISABLED: deferred IPVLAN broke forwarding
-			}
-			// XSK already bound — fall through to reconcile.
-		}
-		if err := ensureFabricIPVLAN(parentLinux, fabLinux, addrs); err != nil {
-			// Fabric overlay is critical for cluster heartbeat and VRRP.
-			// Retry up to 5 times with 1s delay — the parent interface
-			// might not be ready yet after a power cycle.
-			var retryErr error
-			for retry := 0; retry < 5; retry++ {
-				time.Sleep(time.Second)
-				slog.Info("retrying fabric IPVLAN creation",
-					"parent", parentLinux, "name", fabLinux, "attempt", retry+2)
-				retryErr = ensureFabricIPVLAN(parentLinux, fabLinux, addrs)
-				if retryErr == nil {
-					break
-				}
-			}
-			if retryErr != nil {
-				slog.Error("CRITICAL: fabric IPVLAN creation failed after retries — cluster heartbeat will not work",
-					"parent", parentLinux, "name", fabLinux, "err", retryErr)
-			}
-			continue
-		}
-	}
-	// Register deferred IPVLAN creation callback on the userspace manager.
-	if len(deferredOverlays) > 0 {
-		if um, ok := d.dp.(*dpuserspace.Manager); ok {
-			um.OnXSKBound = func() {
-				for _, ov := range deferredOverlays {
-					slog.Info("XSK bound — creating deferred fabric IPVLAN",
-						"parent", ov.parent, "name", ov.name)
-					if err := ensureFabricIPVLAN(ov.parent, ov.name, ov.addrs); err != nil {
-						slog.Error("deferred fabric IPVLAN creation failed",
-							"parent", ov.parent, "name", ov.name, "err", err)
-					}
-				}
-			}
-		}
-	}
-	// Clean up stale fabric IPVLAN overlays not in current config (#128).
-	for _, name := range []string{"fab0", "fab1"} {
-		if activeFabricOverlays[name] {
-			continue
-		}
-		if link, err := netlink.LinkByName(name); err == nil {
-			if _, ok := link.(*netlink.IPVlan); ok {
-				netlink.LinkDel(link)
-				slog.Info("removed stale fabric IPVLAN", "name", name)
-			}
-		}
-	}
-
-	// 1.9. Pre-check: will RETH MAC programming require a link cycle?
-	// If yes, tell the userspace DP to skip initial worker startup during
-	// Compile(). Workers will be started by NotifyLinkCycle() after MAC
-	// programming is done. This avoids the double-bind that causes EBUSY
-	// on mlx5 zero-copy queues.
-	rethMACPending := false
-	if d.cluster != nil && cfg.Chassis.Cluster != nil && d.dp != nil {
-		cc := cfg.Chassis.Cluster
-		for rethName, physName := range cfg.RethToPhysical() {
-			rethCfg, ok := cfg.Interfaces.Interfaces[rethName]
-			if !ok || rethCfg.RedundancyGroup <= 0 {
-				continue
-			}
-			linuxName := config.LinuxIfName(physName)
-			link, err := netlink.LinkByName(linuxName)
-			if err != nil {
-				continue
-			}
-			mac := cluster.RethMAC(cc.ClusterID, rethCfg.RedundancyGroup, cc.NodeID)
-			if !bytes.Equal(link.Attrs().HardwareAddr, mac) {
-				rethMACPending = true
-				break
-			}
-		}
-		if rethMACPending {
-			type deferSetter interface{ SetDeferWorkers(bool) }
-			if ds, ok := d.dp.(deferSetter); ok {
-				ds.SetDeferWorkers(true)
-			}
-		}
-	}
-
-	// 2. Compile eBPF dataplane
-	var compileResult *dataplane.CompileResult
-	if d.dp != nil {
-		var err error
-		if compileResult, err = d.dp.Compile(cfg); err != nil {
-			d.recordCompileFailure(err)
-		} else {
-			d.recordCompileSuccess()
-		}
-	}
-
-	// Clear defer flag after Compile so subsequent recompiles (where MAC
-	// is already set) don't skip workers.
-	if rethMACPending {
-		type deferSetter interface{ SetDeferWorkers(bool) }
-		if ds, ok := d.dp.(deferSetter); ok {
-			ds.SetDeferWorkers(false)
-		}
-	}
-
-	// 2.1. Wire aggressive session aging config to GC.
-	if d.gc != nil {
-		d.gc.SetAgingConfig(
-			cfg.Security.Flow.AgingEarlyAgeout,
-			cfg.Security.Flow.AgingHighWatermark,
-			cfg.Security.Flow.AgingLowWatermark,
-		)
-
-		// Enable per-IP session counting if any screen profile has session limits.
-		sessionLimitEnabled := false
-		for _, sp := range cfg.Security.Screen {
-			if sp.LimitSession.SourceIPBased > 0 || sp.LimitSession.DestinationIPBased > 0 {
-				sessionLimitEnabled = true
-				break
-			}
-		}
-		d.gc.SetSessionLimitEnabled(sessionLimitEnabled)
-	}
-
-	// 2.2. Build zone→RG map for per-RG session sync.
-	if d.sessionSync != nil && compileResult != nil {
-		d.sessionSync.SetZoneRGMap(buildZoneRGMap(cfg, compileResult.ZoneIDs))
-	}
-
-	// 2.5. Write systemd-networkd config for managed interfaces
-	if d.networkd != nil && compileResult != nil && len(compileResult.ManagedInterfaces) > 0 {
-		if err := d.networkd.Apply(compileResult.ManagedInterfaces); err != nil {
-			slog.Warn("failed to apply networkd config", "err", err)
-		}
-	}
-
-	// 2.6. Program deterministic virtual MACs on RETH member interfaces.
-	// Each node gets a per-node MAC (02:bf:72:CC:RR:NN) to avoid FDB conflicts
-	// when both nodes' members are on the same L2 domain. VRRP + gratuitous NA
-	// handle failover; RA goodbye packets handle IPv6 default gateway transitions.
-	// Must run AFTER networkd.Apply() so .link renames are applied first.
-	needLinkCycleRecovery := false
-	if d.cluster != nil && cfg.Chassis.Cluster != nil {
-		cc := cfg.Chassis.Cluster
-		rethToPhys := cfg.RethToPhysical()
-
-		// PrepareLinkCycle is called on-demand after programRethMAC reports
-		// an actual link DOWN/UP cycle. Most drivers (mlx5, virtio) support
-		// IFF_LIVE_ADDR_CHANGE so no cycle is needed and workers keep running.
-
-		for rethName, physName := range rethToPhys {
-			rethCfg, ok := cfg.Interfaces.Interfaces[rethName]
-			if !ok || rethCfg.RedundancyGroup <= 0 {
-				continue
-			}
-			linuxName := config.LinuxIfName(physName)
-			// If the interface doesn't exist under its config name,
-			// find it by RETH virtual MAC and rename it.
-			if _, err := netlink.LinkByName(linuxName); err != nil {
-				mac := cluster.RethMAC(cc.ClusterID, rethCfg.RedundancyGroup, cc.NodeID)
-				if oldName := renameRethMember(linuxName, mac); oldName != "" {
-					slog.Info("renamed RETH member interface",
-						"from", oldName, "to", linuxName)
-					fixRethLinkFile(linuxName, oldName)
-				}
-			}
-			// Ensure the .link file uses OriginalName= (not MACAddress=)
-			// for stable matching across reboots. The bootstrap .link
-			// files may use MACAddress= which breaks after virtual MAC
-			// programming — the interface reboots with physical MAC but
-			// the MACAddress= line might reference the wrong one.
-			ensureRethLinkOriginalName(linuxName)
-			// Disable DAD — virtual MAC may still collide with peer on
-			// some deployments; disable to avoid DAD failures.
-			dadPath := fmt.Sprintf("/proc/sys/net/ipv6/conf/%s/accept_dad", linuxName)
-			os.WriteFile(dadPath, []byte("0"), 0644)
-			// Suppress auto link-local generation on RETH member interfaces.
-			// The virtual MAC triggers a kernel-generated link-local (fe80::...)
-			// which causes continuous MLDv2 multicast reports on the L2 segment.
-			// VIPs are managed explicitly; auto link-locals are unnecessary.
-			addrGenPath := fmt.Sprintf("/proc/sys/net/ipv6/conf/%s/addr_gen_mode", linuxName)
-			os.WriteFile(addrGenPath, []byte("1"), 0644)
-			mac := cluster.RethMAC(cc.ClusterID, rethCfg.RedundancyGroup, cc.NodeID)
-			linkCycled, err := programRethMAC(linuxName, mac)
-			if err != nil {
-				slog.Warn("failed to set RETH MAC", "iface", linuxName, "mac", mac, "err", err)
-			}
-			if linkCycled && !needLinkCycleRecovery {
-				// First link cycle — stop workers NOW (they may have
-				// been accessing UMEM during the DOWN/UP). The rebind
-				// in NotifyLinkCycle will restart them.
-				if d.dp != nil {
-					type linkCyclePreparer interface{ PrepareLinkCycle() }
-					if preparer, ok := d.dp.(linkCyclePreparer); ok {
-						slog.Info("userspace: stopping workers after RETH MAC link cycle")
-						preparer.PrepareLinkCycle()
-					}
-				}
-			}
-			needLinkCycleRecovery = needLinkCycleRecovery || linkCycled
-			clearDadFailed(linuxName)
-			removeAutoLinkLocal(linuxName)
-			// Re-add link-local if this parent interface has IPv6 on unit 0.
-			// NDP Neighbor Solicitation requires a link-local source address.
-			if rethUnitHasIPv6(rethCfg, 0) {
-				ensureRethLinkLocal(linuxName)
-			}
-
-			// Re-disable VLAN RX offload after MAC programming.
-			// The iavf VF driver resets ethtool features (including
-			// rx-vlan-offload) during the link down/up cycle that
-			// programRethMAC requires. Without this, XDP cannot see
-			// VLAN tags in the packet data and drops VLAN traffic.
-			if out, err := exec.Command("ethtool", "-K", linuxName, "rxvlan", "off").CombinedOutput(); err != nil {
-				slog.Warn("failed to re-disable rxvlan after RETH MAC",
-					"interface", linuxName, "err", err, "output", strings.TrimSpace(string(out)))
-			} else {
-				slog.Info("re-disabled VLAN RX offload after RETH MAC", "interface", linuxName)
-			}
-
-			// Propagate MAC change to VLAN sub-interfaces.
-			// Linux VLAN sub-interfaces don't always inherit the
-			// parent's MAC change after link down/up.
-			if parentLink, err := netlink.LinkByName(linuxName); err == nil {
-				parentIdx := parentLink.Attrs().Index
-				links, _ := netlink.LinkList()
-				for _, l := range links {
-					if l.Attrs().ParentIndex != parentIdx {
-						continue
-					}
-					subName := l.Attrs().Name
-					// Suppress auto link-local on VLAN sub-interfaces too.
-					subAddrGen := fmt.Sprintf("/proc/sys/net/ipv6/conf/%s/addr_gen_mode", subName)
-					os.WriteFile(subAddrGen, []byte("1"), 0644)
-					if !bytes.Equal(l.Attrs().HardwareAddr, mac) {
-						if err := netlink.LinkSetHardwareAddr(l, mac); err != nil {
-							slog.Warn("failed to propagate MAC to VLAN sub-interface",
-								"iface", subName, "err", err)
-						} else {
-							slog.Info("propagated RETH MAC to VLAN sub-interface",
-								"iface", subName, "mac", mac)
-						}
-					}
-					removeAutoLinkLocal(subName)
-					// Re-add link-local if this VLAN sub-interface has IPv6.
-					// Extract VLAN ID from sub-interface name (e.g. "ge-7-0-1.100").
-					if dotIdx := strings.LastIndex(subName, "."); dotIdx >= 0 {
-						if vid, err := strconv.Atoi(subName[dotIdx+1:]); err == nil {
-							if rethUnitHasIPv6(rethCfg, vid) {
-								ensureRethLinkLocal(subName)
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-
-	// 2.6b. Reconcile VRRP VIPs and stable link-locals after RETH MAC
-	// programming. Only needed when programRethMAC had to bring the
-	// interface DOWN/UP (link cycle), which removes all addresses
-	// including VRRP VIPs and stable link-locals.
-	if needLinkCycleRecovery && d.isNoRethVRRP() {
-		// Direct mode: re-add VIPs + stable link-locals for each RG
-		// where we are primary.
-		if d.cluster != nil {
-			for _, rg := range cfg.Chassis.Cluster.RedundancyGroups {
-				if d.cluster.IsLocalPrimary(rg.ID) {
-					d.directAddVIPs(rg.ID)
-					d.addStableRethLinkLocal(rg.ID)
-					d.scheduleDirectAnnounce(rg.ID, "link-cycle-recovery")
-				}
-			}
-		}
-	} else if needLinkCycleRecovery && d.vrrpMgr != nil {
-		d.vrrpMgr.ReconcileVIPs()
-		// Re-add stable link-locals for active RGs after MAC bounce.
-		if d.cluster != nil && cfg.Chassis.Cluster != nil {
-			for _, rg := range cfg.Chassis.Cluster.RedundancyGroups {
-				s := d.getOrCreateRGState(rg.ID)
-				if s.IsActive() {
-					d.addStableRethLinkLocal(rg.ID)
-				}
-			}
-		}
-	}
-
-	// 2.6b2. Rebind AF_XDP sockets after RETH MAC programming.
-	// Only needed when PrepareLinkCycle was called (macChangeNeeded=true
-	// or rethMACPending=true). Calling NotifyLinkCycle without a prior
-	// PrepareLinkCycle causes a spurious rebind that gets EBUSY on mlx5
-	// zero-copy queues because the first bind is still in progress.
-	if d.dp != nil && needLinkCycleRecovery {
-		// Actual link DOWN/UP occurred — old XSK sockets are dead.
-		// Rebind to create fresh sockets on the reinitialized queues.
-		d.dp.NotifyLinkCycle()
-		if d.ra != nil {
-			d.ra.ResendBurst()
-		}
-	} else if d.dp != nil && rethMACPending && !needLinkCycleRecovery {
-		// MAC set live (no link cycle) but workers were deferred.
-		// Trigger a re-Compile to start workers with the now-correct MAC.
-		// This is cheaper than NotifyLinkCycle (no stop_workers/rebind).
-		if _, err := d.dp.Compile(cfg); err != nil {
-			slog.Warn("failed to re-compile after deferred MAC", "err", err)
-		}
-	}
-
-	// NOTE: stable link-local cleanup for secondary RGs is handled by
-	// the reconcile loop (reconcileRGState) after election settles,
-	// not here — we don't know who's primary during config apply.
-
-	// 2.6c. Reconcile proxy ARP entries for NAT addresses.
-	if len(cfg.Security.NAT.ProxyARP) > 0 {
-		ifaceMap := make(map[string]int)
-		rethToPhys := cfg.RethToPhysical()
-		for _, entry := range cfg.Security.NAT.ProxyARP {
-			parts := strings.SplitN(entry.Interface, ".", 2)
-			baseName := parts[0]
-			if phys, ok := rethToPhys[baseName]; ok {
-				baseName = phys
-			}
-			linuxName := config.LinuxIfName(baseName)
-			if _, ok := ifaceMap[entry.Interface]; ok {
-				continue
-			}
-			iface, err := net.InterfaceByName(linuxName)
-			if err != nil {
-				slog.Warn("proxy-arp: interface not found", "iface", entry.Interface, "linux", linuxName, "err", err)
-				continue
-			}
-			ifaceMap[entry.Interface] = iface.Index
-		}
-		added, err := dataplane.ReconcileProxyARP(cfg, ifaceMap)
-		if err != nil {
-			slog.Warn("failed to reconcile proxy ARP", "err", err)
-		}
-		for _, a := range added {
-			if a.Iface != "" {
-				if err := cluster.SendGratuitousARP(a.Iface, a.IP, 1); err != nil {
-					slog.Warn("proxy-arp: GARP failed", "ip", a.IP, "iface", a.Iface, "err", err)
-				}
-			}
-		}
-	}
-
-	// 2.7. Re-bind management VRF interfaces after networkd.Apply().
-	// networkctl reconfigure strips VRF master bindings because networkd
-	// considers the daemon-created vrf-mgmt device "unmanaged" and ignores
-	// the VRF= directive. Re-bind here to restore VRF membership.
-	if d.routing != nil && d.mgmtVRFInterfaces != nil {
-		for ifName := range d.mgmtVRFInterfaces {
-			if err := d.routing.BindInterfaceToVRF(ifName, "mgmt"); err != nil {
-				slog.Warn("failed to re-bind interface to management VRF",
-					"interface", ifName, "err", err)
-			}
-		}
-		// Restart heartbeat after VRF rebind — networkd reconfigure moves
-		// the control interface (em0) out of vrf-mgmt temporarily, which
-		// invalidates the heartbeat UDP sockets. Without this restart,
-		// the recovering node stops receiving peer heartbeats and declares
-		// split-brain after the grace period expires.
-		if d.cluster != nil {
-			d.cluster.RestartHeartbeat()
-		}
-	}
-
-	// 3. Apply all routes + dynamic protocols via FRR
-	if d.frr != nil {
-		// Collect interface bandwidths and point-to-point flags for FRR.
-		ifaceBandwidths := make(map[string]uint64)
-		ifaceP2P := make(map[string]bool)
-		for name, ifc := range cfg.Interfaces.Interfaces {
-			linuxName := config.LinuxIfName(name)
-			if ifc.Bandwidth > 0 {
-				ifaceBandwidths[linuxName] = ifc.Bandwidth
-			}
-			for _, unit := range ifc.Units {
-				if unit.PointToPoint {
-					ifaceP2P[linuxName] = true
-				}
-			}
-		}
-
-		fc := &frr.FullConfig{
-			OSPF:                  cfg.Protocols.OSPF,
-			OSPFv3:                cfg.Protocols.OSPFv3,
-			BGP:                   cfg.Protocols.BGP,
-			RIP:                   cfg.Protocols.RIP,
-			ISIS:                  cfg.Protocols.ISIS,
-			StaticRoutes:          cfg.RoutingOptions.StaticRoutes,
-			Inet6StaticRoutes:     cfg.RoutingOptions.Inet6StaticRoutes,
-			GenerateRoutes:        cfg.RoutingOptions.GenerateRoutes,
-			DHCPRoutes:            d.collectDHCPRoutes(),
-			PolicyOptions:         &cfg.PolicyOptions,
-			ForwardingTableExport: cfg.RoutingOptions.ForwardingTableExport,
-			BackupRouter:          cfg.System.BackupRouter,
-			BackupRouterDst:       cfg.System.BackupRouterDst,
-			InterfaceBandwidths:   ifaceBandwidths,
-			InterfacePointToPoint: ifaceP2P,
-			RethMap:               cfg.RethToPhysical(),
-			IPv6NextHopInterfaces: inferIPv6StaticNextHopInterfaces(cfg),
-			ClusterMode:           d.cluster != nil,
-		}
-		for _, ri := range cfg.RoutingInstances {
-			vrfName := "vrf-" + ri.Name
-			if ri.InstanceType == "forwarding" {
-				vrfName = "" // forwarding instances use the default table
-			}
-			fc.Instances = append(fc.Instances, frr.InstanceConfig{
-				VRFName:           vrfName,
-				OSPF:              ri.OSPF,
-				OSPFv3:            ri.OSPFv3,
-				BGP:               ri.BGP,
-				RIP:               ri.RIP,
-				ISIS:              ri.ISIS,
-				StaticRoutes:      ri.StaticRoutes,
-				Inet6StaticRoutes: ri.Inet6StaticRoutes,
-			})
-		}
-		if err := d.frr.ApplyFull(fc); err != nil {
-			slog.Warn("failed to apply FRR config", "err", err)
-		}
-
-		// Set L4 ECMP hash when consistent-hash is configured.
-		if fc.ConsistentHash {
-			path := "/proc/sys/net/ipv4/fib_multipath_hash_policy"
-			current, _ := os.ReadFile(path)
-			if strings.TrimSpace(string(current)) != "1" {
-				if err := os.WriteFile(path, []byte("1\n"), 0644); err != nil {
-					slog.Warn("failed to set fib_multipath_hash_policy", "err", err)
-				} else {
-					slog.Info("enabled L4 ECMP hashing (consistent-hash)")
-				}
-			}
-		}
-	}
-
-	// 3b. Apply next-table policy routing rules (ip rule)
-	if d.routing != nil {
-		// Collect all static routes from main + per-rib
-		allRoutes := append(cfg.RoutingOptions.StaticRoutes, cfg.RoutingOptions.Inet6StaticRoutes...)
-		if err := d.routing.ApplyNextTableRules(allRoutes, cfg.RoutingInstances); err != nil {
-			slog.Warn("failed to apply next-table rules", "err", err)
-		}
-	}
-
-	// 3c. Apply rib-group route leaking rules (ip rule)
-	if d.routing != nil && len(cfg.RoutingOptions.RibGroups) > 0 {
-		if err := d.routing.ApplyRibGroupRules(cfg.RoutingOptions.RibGroups, cfg.RoutingInstances); err != nil {
-			slog.Warn("failed to apply rib-group rules", "err", err)
-		}
-	}
-
-	// 3d. Apply policy-based routing rules (ip rule) for firewall filter routing-instance
-	if d.routing != nil {
-		pbrRules := routing.BuildPBRRules(&cfg.Firewall, cfg.RoutingInstances)
-		if err := d.routing.ApplyPBRRules(pbrRules); err != nil {
-			slog.Warn("failed to apply PBR rules", "err", err)
-		}
-	}
-
-	// 4. Proactive neighbor resolution for all known next-hops/gateways.
-	// This ensures bpf_fib_lookup returns SUCCESS (with valid MACs)
-	// instead of NO_NEIGH for the first forwarded packet.
-	// In cluster mode, skip here — RETH VIPs are not yet installed (VRRP
-	// hasn't become MASTER), so RouteGet() for WAN next-hops may fail.
-	// resolveNeighbors() is triggered on VRRP MASTER in watchVRRPEvents.
-	if cfg.Chassis.Cluster == nil {
-		d.resolveNeighbors(cfg)
-	}
-
-	// 5. Apply RA config (Router Advertisements)
-	// In cluster mode, RA/kea are managed by watchVRRPEvents — only
-	// the MASTER runs these services to prevent dual-RA / dual-DHCP.
-	// The VRRP event fires shortly after startup and calls applyRethServices().
-	isCluster := cfg.Chassis.Cluster != nil
-	raConfigs := d.buildRAConfigs(cfg)
-	if !isCluster {
-		if d.ra != nil && len(raConfigs) > 0 {
-			if err := d.ra.Apply(raConfigs); err != nil {
-				slog.Warn("failed to apply RA config", "err", err)
-			}
-		} else if d.ra != nil {
-			// No RA configs — clear any previous RA senders.
-			if err := d.ra.Clear(); err != nil {
-				slog.Warn("failed to clear RA config", "err", err)
-			}
-		}
-	}
-	// Cluster startup: goodbye RAs for stale routes are handled by the
-	// reconcile loop (reconcileRGState) after VRRP election settles.
-	// Each RETH node has a different virtual MAC (hence different
-	// link-local), so both nodes appear as separate routers to hosts.
-	// Only the primary sends RAs (via applyRethServicesForRG on MASTER);
-	// the reconcile loop sends goodbye RAs for inactive RGs.
-	//
-	// Stable link-local cleanup: handled by reconcile after election.
-
-	// 6. Apply IPsec config
-	// Always call Apply so stale swanctl config is removed when VPNs are
-	// deleted from config.
-	if d.ipsec != nil {
-		if err := d.ipsec.Apply(ipsec.PrepareConfig(cfg)); err != nil {
-			slog.Warn("failed to apply IPsec config", "err", err)
-		}
-	}
-
-	// 7. Apply DHCP server config (Kea DHCPv4 + DHCPv6)
-	// In cluster mode, deferred to VRRP MASTER transition.
-	if !isCluster && d.dhcpServer != nil && (cfg.System.DHCPServer.DHCPLocalServer != nil || cfg.System.DHCPServer.DHCPv6LocalServer != nil) {
-		// Resolve RETH interface names for Kea (needs real Linux names)
-		resolveDHCPRethInterfaces(&cfg.System.DHCPServer, cfg)
-		if err := d.dhcpServer.Apply(&cfg.System.DHCPServer); err != nil {
-			slog.Warn("failed to apply DHCP server config", "err", err)
-		}
-	}
-
-	// 8. Apply VRRP config — merge user VRRP + RETH VRRP instances
-	vrrpInstances := vrrp.CollectInstances(cfg)
-	if d.cluster != nil {
-		localPri := d.cluster.LocalPriorities()
-		vrrpInstances = append(vrrpInstances, vrrp.CollectRethInstances(cfg, localPri)...)
-	}
-	if err := d.vrrpMgr.UpdateInstances(vrrpInstances); err != nil {
-		slog.Warn("failed to update VRRP instances", "err", err)
-	}
-
-	// 9. Apply system DNS and NTP configuration
-	d.applySystemDNS(cfg)
-	d.applySystemNTP(cfg)
-	d.applyDNSService(cfg)
-
-	// 9.5. Apply system hostname, timezone, and kernel tuning
-	d.applyHostname(cfg)
-	d.applyTimezone(cfg)
-	d.applyKernelTuning(cfg)
-	d.applyLo0Filter(cfg)
-
-	// 9.6. Write SSH known hosts file
-	d.applySSHKnownHosts(cfg)
-
-	// 10. Apply system syslog forwarding
-	d.applySystemSyslog(cfg)
-
-	// 11. Apply system login users (create OS accounts, SSH keys)
-	d.applySystemLogin(cfg)
-
-	// 12. Apply SSH service configuration (root-login)
-	d.applySSHConfig(cfg)
-
-	// 13. Apply root authentication (encrypted-password + SSH keys)
-	d.applyRootAuth(cfg)
-
-	// 14. Apply syslog file destinations (rsyslog configs)
-	d.applySyslogFiles(cfg)
-
-	// 14b. Update security log syslog clients + zone name mapping
-	if d.eventReader != nil {
-		d.applySyslogConfig(d.eventReader, cfg)
-	}
-
-	// 15. Archive config to remote sites if transfer-on-commit is enabled
-	d.archiveConfig(cfg)
-
-	// 15b. Configure local archival settings for auto-archive on commit
-	if cfg.System.Archival != nil {
-		dir := cfg.System.Archival.ArchiveDir
-		if dir == "" {
-			dir = "/var/lib/xpf/archive"
-		}
-		max := cfg.System.Archival.MaxArchives
-		if max <= 0 {
-			max = 10
-		}
-		d.store.SetArchiveConfig(dir, max)
-	} else {
-		d.store.SetArchiveConfig("", 0)
-	}
-
-	// 16. Update flow traceoptions (trace file + filters)
-	d.updateFlowTrace(cfg)
-
-	// 17. Update event-options policies (RPM-driven failover)
-	if d.eventEngine != nil {
-		d.eventEngine.Apply(cfg.EventOptions)
-	}
-
-	// 18. Update chassis cluster interface monitors
-	if d.routing != nil && cfg.Chassis.Cluster != nil &&
-		len(cfg.Chassis.Cluster.RedundancyGroups) > 0 {
-		d.routing.ApplyInterfaceMonitors(cfg.Chassis.Cluster.RedundancyGroups)
-	}
-
-	// 19. Update chassis cluster state machine
-	if d.cluster != nil && cfg.Chassis.Cluster != nil {
-		d.cluster.UpdateConfig(cfg.Chassis.Cluster)
-		// Feed interface monitor statuses into cluster weight calculation
-		if d.routing != nil {
-			monStatuses := d.routing.InterfaceMonitorStatuses()
-			for rgID, statuses := range monStatuses {
-				for _, st := range statuses {
-					d.cluster.SetMonitorWeight(rgID, st.Interface, !st.Up, st.Weight)
-				}
-			}
-		}
-
-		// RETH GARP is handled by native VRRP (VRRP-backed RETH).
-		// No manual GARP registration needed.
-	}
-
-	// 20. Detect cluster transport config changes and restart comms (#87).
-	// Only restart if comms were previously started (activeClusterTransport
-	// is non-zero) and the new config differs.
-	if d.cluster != nil && d.daemonCtx != nil {
-		newTransport := clusterTransportFromConfig(cfg)
-		if d.activeClusterTransport != (clusterTransportKey{}) && newTransport != d.activeClusterTransport {
-			slog.Info("cluster: transport config changed, restarting comms",
-				"old_control", d.activeClusterTransport.ControlInterface,
-				"new_control", newTransport.ControlInterface,
-				"old_peer", d.activeClusterTransport.PeerAddress,
-				"new_peer", newTransport.PeerAddress,
-				"old_fabric", d.activeClusterTransport.FabricInterface,
-				"new_fabric", newTransport.FabricInterface,
-				"old_fabric_peer", d.activeClusterTransport.FabricPeerAddress,
-				"new_fabric_peer", newTransport.FabricPeerAddress)
-			d.stopClusterComms()
-			d.startClusterComms(d.daemonCtx)
-		}
-	}
-
-	// 21. Re-apply D3 RSS indirection on config change (#797 HIGH #2).
-	// Worker count can change via commit (e.g. `set system dataplane
-	// workers 6`), and the D3 disable knob can flip; either requires
-	// re-running the reshape (or restore) against the current HW state.
-	// Idempotent: matching tables skip the write. Non-mlx5 interfaces
-	// are skipped at the per-interface guard. The allowlist is
-	// recomputed from the *new* compiled config so interface-set
-	// changes (added/removed zoned mlx5 interfaces, fabric interface
-	// changes) take effect on the same commit.
-	if !d.opts.NoDataplane {
-		rssEnabled := true
-		workers := 0
-		var rssAllowed []string
-		// #801: mirror the startup site so a commit that changes any
-		// of the Step-0 knobs takes effect without a restart.
-		var (
-			governor          string
-			netdevBudget      int
-			coalesceEnable    bool
-			coalesceRX        int
-			coalesceTX        int
-			userspaceDP       bool
-			coalesceExplicit  bool
-			claimHostTunables bool
-		)
-		if cfg.System.DataplaneType == "userspace" && cfg.System.UserspaceDataplane != nil {
-			userspaceDP = true
-			workers = cfg.System.UserspaceDataplane.Workers
-			if cfg.System.UserspaceDataplane.RSSIndirectionDisabled {
-				rssEnabled = false
-			}
-			rssAllowed = dpuserspace.UserspaceBoundLinuxInterfaces(cfg)
-			claimHostTunables = cfg.System.UserspaceDataplane.ClaimHostTunables
-			governor = cfg.System.UserspaceDataplane.CPUGovernor
-			netdevBudget = cfg.System.UserspaceDataplane.NetdevBudget
-			coalesceExplicit = cfg.System.UserspaceDataplane.CoalescenceAdaptiveExplicit
-			if coalesceExplicit &&
-				!cfg.System.UserspaceDataplane.CoalescenceAdaptiveDisabled {
-				coalesceEnable = true
-			}
-			coalesceRX = cfg.System.UserspaceDataplane.CoalescenceRXUsecs
-			coalesceTX = cfg.System.UserspaceDataplane.CoalescenceTXUsecs
-		}
-		reapplyRSSIndirection(rssEnabled, workers, rssAllowed)
-		// #801 B1 + B2: opt-in gate + restore-on-disable.
-		d.applyStep0Tunables(userspaceDP, claimHostTunables, governor, netdevBudget,
-			coalesceExplicit, coalesceEnable, coalesceRX, coalesceTX, rssAllowed)
-	}
-}
-
-// buildRAConfigs merges static RA configs from the Junos config with
-// PD-derived prefixes from DHCPv6 prefix delegation.
-func (d *Daemon) buildRAConfigs(cfg *config.Config) []*config.RAInterfaceConfig {
-	// Start with static RA configs from the configuration.
-	raByIface := make(map[string]*config.RAInterfaceConfig)
-	var result []*config.RAInterfaceConfig
-	for _, ra := range cfg.Protocols.RouterAdvertisement {
-		clone := cloneRAInterfaceConfig(ra)
-		raByIface[clone.Interface] = clone
-		result = append(result, clone)
-	}
-
-	if d.dhcp != nil {
-		// Merge PD-derived prefixes from DHCPv6 clients.
-		for _, mapping := range d.dhcp.DelegatedPrefixesForRA() {
-			subPrefix := dhcp.DeriveSubPrefix(mapping.Prefix, mapping.SubPrefLen)
-			if !subPrefix.IsValid() {
-				slog.Warn("DHCPv6 PD: invalid sub-prefix derivation",
-					"delegated", mapping.Prefix, "sub_len", mapping.SubPrefLen)
-				continue
-			}
-
-			pfx := &config.RAPrefix{
-				Prefix:     subPrefix.String(),
-				OnLink:     true,
-				Autonomous: true,
-			}
-			if mapping.ValidLifetime > 0 {
-				pfx.ValidLifetime = int(mapping.ValidLifetime.Seconds())
-			}
-			if mapping.PreferredLifetime > 0 {
-				pfx.PreferredLife = int(mapping.PreferredLifetime.Seconds())
-			}
-
-			if existing, ok := raByIface[mapping.RAIface]; ok {
-				// Append prefix to existing RA interface config.
-				existing.Prefixes = append(existing.Prefixes, pfx)
-			} else {
-				// Create a new RA interface config for this downstream interface.
-				ra := &config.RAInterfaceConfig{
-					Interface: mapping.RAIface,
-					Prefixes:  []*config.RAPrefix{pfx},
-				}
-				raByIface[mapping.RAIface] = ra
-				result = append(result, ra)
-			}
-
-			slog.Info("DHCPv6 PD: advertising prefix via RA",
-				"prefix", subPrefix, "interface", mapping.RAIface,
-				"delegated_from", mapping.Interface)
-		}
-	}
-
-	// Detect explicitly configured link-local addresses on RA interfaces.
-	// If a user configures e.g. fe80::face/64 on a RETH interface, the RA
-	// sender should bind to that address instead of auto-selecting one.
-	for _, ra := range result {
-		if ifc, ok := cfg.Interfaces.Interfaces[ra.Interface]; ok {
-			if unit, ok := ifc.Units[0]; ok {
-				for _, addr := range unit.Addresses {
-					ip, _, err := net.ParseCIDR(addr)
-					if err != nil {
-						continue
-					}
-					if ip.IsLinkLocalUnicast() && ip.To4() == nil {
-						ra.SourceLinkLocal = ip.String()
-						break
-					}
-				}
-			}
-			if ra.SourceLinkLocal == "" && cfg.Chassis.Cluster != nil && ifc.RedundancyGroup != 0 {
-				// RETH HA startup installs a stable router link-local on the active
-				// member. Bind RA to that address so the sender does not auto-pick a
-				// transient EUI-64 link-local which can later be removed by HA reconcile.
-				ra.SourceLinkLocal = cluster.StableRethLinkLocal(
-					cfg.Chassis.Cluster.ClusterID,
-					ifc.RedundancyGroup,
-				).String()
-			}
-		}
-	}
-
-	// Resolve RETH interface names for RA senders (needs real Linux names).
-	for _, ra := range result {
-		ra.Interface = config.LinuxIfName(cfg.ResolveReth(ra.Interface))
-	}
-
-	return result
-}
-
-func cloneRAInterfaceConfig(src *config.RAInterfaceConfig) *config.RAInterfaceConfig {
-	if src == nil {
-		return nil
-	}
-	clone := *src
-	if len(src.DNSServers) > 0 {
-		clone.DNSServers = append([]string(nil), src.DNSServers...)
-	}
-	if len(src.Prefixes) > 0 {
-		clone.Prefixes = make([]*config.RAPrefix, 0, len(src.Prefixes))
-		for _, pfx := range src.Prefixes {
-			if pfx == nil {
-				clone.Prefixes = append(clone.Prefixes, nil)
-				continue
-			}
-			pfxClone := *pfx
-			clone.Prefixes = append(clone.Prefixes, &pfxClone)
-		}
-	}
-	return &clone
-}
-
-// startDHCPClients iterates the config and starts DHCP/DHCPv6 clients
-// for interfaces that have family inet { dhcp; } or family inet6 { dhcpv6; }.
-func (d *Daemon) startDHCPClients(ctx context.Context, cfg *config.Config) {
-	// Check if any interface needs DHCP
-	needsDHCP := false
-	for _, ifc := range cfg.Interfaces.Interfaces {
-		for _, unit := range ifc.Units {
-			if unit.DHCP || unit.DHCPv6 {
-				needsDHCP = true
-				break
-			}
-		}
-	}
-	if !needsDHCP {
-		return
-	}
-
-	// State dir for DUID persistence — same directory as config file
-	stateDir := filepath.Dir(d.opts.ConfigFile)
-
-	dm, err := dhcp.New(stateDir, func() {
-		// Full recompile is safe: heartbeat sockets survive VRF rebind
-		// (RestartHeartbeat), RETH MAC is set live (no XSK rebind), and
-		// BPF compile skips reconcile when the binding plan is unchanged.
-		if activeCfg := d.store.ActiveConfig(); activeCfg != nil {
-			if d.dhcpLeaseChangeRequiresRecompile(activeCfg) {
-				slog.Info("DHCP address changed, recompiling dataplane")
-				d.applyConfig(activeCfg)
-			} else {
-				slog.Info("DHCP address changed on management-only interface, refreshing management routes")
-				d.applyMgmtVRFRoutes()
-			}
-		}
-	})
-	if err != nil {
-		slog.Warn("failed to create DHCP manager", "err", err)
-		return
-	}
-	d.dhcp = dm
-
-	for ifName, ifc := range cfg.Interfaces.Interfaces {
-		for _, unit := range ifc.Units {
-			// Use VLAN sub-interface name when unit has a VLAN ID
-			dhcpIface := config.LinuxIfName(ifName)
-			if unit.VlanID > 0 {
-				dhcpIface = fmt.Sprintf("%s.%d", dhcpIface, unit.VlanID)
-			}
-			if unit.DHCP {
-				if unit.DHCPOptions != nil {
-					dm.SetDHCPv4Options(dhcpIface, &dhcp.DHCPv4Options{
-						LeaseTime:              unit.DHCPOptions.LeaseTime,
-						RetransmissionAttempt:  unit.DHCPOptions.RetransmissionAttempt,
-						RetransmissionInterval: unit.DHCPOptions.RetransmissionInterval,
-						ForceDiscover:          unit.DHCPOptions.ForceDiscover,
-					})
-				}
-				slog.Info("starting DHCPv4 client", "interface", dhcpIface)
-				dm.Start(ctx, dhcpIface, dhcp.AFInet)
-			}
-			if unit.DHCPv6 {
-				// Configure DUID type from dhcpv6-client stanza
-				if unit.DHCPv6Client != nil && unit.DHCPv6Client.DUIDType != "" {
-					dm.SetDUIDType(dhcpIface, unit.DHCPv6Client.DUIDType)
-				} else {
-					dm.SetDUIDType(dhcpIface, "duid-ll") // default
-				}
-				// Configure DHCPv6 PD and other options
-				if unit.DHCPv6Client != nil {
-					dm.SetDHCPv6Options(dhcpIface, &dhcp.DHCPv6Options{
-						Stateless:  unit.DHCPv6Client.ClientType == "stateless",
-						UpdateDNS:  slices.Contains(unit.DHCPv6Client.ReqOptions, "dns-server"),
-						IATypes:    unit.DHCPv6Client.ClientIATypes,
-						PDPrefLen:  unit.DHCPv6Client.PrefixDelegatingPrefixLen,
-						PDSubLen:   unit.DHCPv6Client.PrefixDelegatingSubPrefLen,
-						ReqOptions: unit.DHCPv6Client.ReqOptions,
-						RAIface:    unit.DHCPv6Client.UpdateRAInterface,
-					})
-				}
-				slog.Info("starting DHCPv6 client", "interface", dhcpIface)
-				dm.Start(ctx, dhcpIface, dhcp.AFInet6)
-			}
-		}
-	}
-}
-
-func (d *Daemon) dhcpLeaseChangeRequiresRecompile(cfg *config.Config) bool {
-	if cfg == nil {
-		return false
-	}
-	// Prefix delegation can affect downstream addressing/RA and still needs
-	// a full re-apply.
-	if d.dhcp != nil && len(d.dhcp.DelegatedPrefixesForRA()) > 0 {
-		return true
-	}
-	// If management VRF bindings are unavailable, stay conservative.
-	if len(d.mgmtVRFInterfaces) == 0 {
-		return true
-	}
-	for ifName, ifc := range cfg.Interfaces.Interfaces {
-		if ifc == nil {
-			continue
-		}
-		for _, unit := range ifc.Units {
-			if unit == nil || (!unit.DHCP && !unit.DHCPv6) {
-				continue
-			}
-			dhcpIface := config.LinuxIfName(ifName)
-			if unit.VlanID > 0 {
-				dhcpIface = fmt.Sprintf("%s.%d", dhcpIface, unit.VlanID)
-			}
-			if !d.mgmtVRFInterfaces[dhcpIface] {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-// resolveJunosIfName converts a Junos-style interface name to its Linux
-// equivalent. It resolves RETH names to their physical members (e.g.
-// reth0.50 → ge-0/0/0.50) and converts Junos slashes to dashes (e.g.
-// ge-0/0/0 → ge-0-0-0).
-func resolveJunosIfName(cfg *config.Config, ifName string) string {
-	return config.LinuxIfName(cfg.ResolveReth(ifName))
-}
-
-func resolveConfigSubnetLinuxName(cfg *config.Config, ip net.IP) (string, string, bool) {
-	if cfg == nil || ip == nil {
-		return "", "", false
-	}
-	for _, ifc := range cfg.Interfaces.Interfaces {
-		if ifc == nil {
-			continue
-		}
-		for unitNum, unit := range ifc.Units {
-			if unit == nil {
-				continue
-			}
-			for _, addrStr := range unit.Addresses {
-				_, ipNet, err := net.ParseCIDR(addrStr)
-				if err != nil {
-					continue
-				}
-				if !ipNet.Contains(ip) {
-					continue
-				}
-				ifName := resolveJunosIfName(cfg, ifc.Name)
-				if unit.VlanID > 0 {
-					ifName = fmt.Sprintf("%s.%d", ifName, unit.VlanID)
-				} else if unitNum != 0 {
-					ifName = fmt.Sprintf("%s.%d", ifName, unitNum)
-				}
-				return ifName, addrStr, true
-			}
-		}
-	}
-	return "", "", false
-}
-
-// stripCIDR removes the /prefix from a CIDR string, returning just the IP.
-func stripCIDR(s string) string {
-	ip, _, err := net.ParseCIDR(s)
-	if err != nil {
-		return s // not CIDR, return as-is
-	}
-	return ip.String()
 }
